### 2020-10-07 #2730 @whedon
[PRE REVIEW]: CoPro: a data-driven model for conflict risk projections
**Submitting author:** @JannisHoch (<a href="http://orcid.org/0000-0003-3570-6436">Jannis Hoch</a>)
**Repository:** <a href="https://github.com/JannisHoch/copro" target ="_blank">https://github.com/JannisHoch/copro</a>
**Version:** v0.0.5
