### 2020-10-15 #2730 @martibosch
Hello @sbenthall,

after going through the README, documentation and example notebooks, I am not sure that the library falls within my expertise (besides my "general" knowledge of the underlying machine learning and GIS involved). Nonetheless, if you do not find other JOSS reviewers that are better suited to the task, I will do my best to review this.

Cheers,
Martí
