### 2020-10-18 #2730 @Zac-HD
@sbenthall - I'd be happy to do a generic code review, but don't feel I know the domain well enough to review the paper.
