### 2020-10-23 #2730 @platipodium
> @platipodium would you be able to review this article?

I am not an expert in conflict or risk modeling, nor machine learning, so I'd like to deflect this to a more knowledgeable reviewer.   May I suggest marcel.nonnenmacher@hzg.de for machine learning and roland.comier@hzg.de for risk assessment.

On thing I noticed: From the existing ReadtheDocs and ReadMe and associated paper, it is still unclear to me, what the software is good for.  An example output/visualization in the JOSS paper and/or documentation would be helpful to find reviewers, and users.
