### 2020-11-04 e9c5a85
updated writing of output files to folder; also in command line script
