+++ scripts/runner.py
@@ -91,7 +91,11 @@ def main(cfg):
         if config.getboolean('general', 'verbose'):
             print('average {0} of run with {1} repetitions is {2:0.3f}'.format(key, config.getint('settings', 'n_runs'), np.mean(out_dict[key])))
 
-    df_hit, gdf_hit = copro.evaluation.polygon_model_accuracy(out_y_df, global_df, out_dir=None)
+    # create accuracy values per polygon and save to output folder
+    df_hit, gdf_hit = copro.evaluation.polygon_model_accuracy(out_y_df, global_df, out_dir)
+
+    # apply k-fold 
+    gdf_CCP = copro.evaluation.calc_kFold_polygon_analysis(out_y_df, global_df, out_dir, k=10)
 
     #- plot distribution of all evaluation metrics
     fig, ax = plt.subplots(1, 1)
