+++ README.rst
@@ -7,8 +7,8 @@ CoPro
 
 Welcome to CoPro, a machine-learning tool for conflict risk projections based on climate, environmental, and societal drivers.
 
-.. image:: https://travis-ci.com/JannisHoch/conflict_model.svg?branch=dev
-    :target: https://travis-ci.com/JannisHoch/conflict_model
+.. image:: https://travis-ci.com/JannisHoch/copro.svg?branch=dev
+    :target: https://travis-ci.com/JannisHoch/copro
 
 .. image:: https://img.shields.io/badge/License-MIT-blue.svg
     :target: https://github.com/JannisHoch/copro/blob/dev/LICENSE
