### 2020-11-06 24e7c22
towards v0.0.6b for making future projections withtout observations
