+++ setup.py
@@ -45,6 +45,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/JannisHoch/copro',
-    version='0.0.5',
+    version='0.0.6b',
     zip_safe=False,
 )
