### 2020-11-06 701febe
updated functions and added code to make predictions
