+++ copro/pipeline.py
@@ -4,7 +4,7 @@ import numpy as np
 import os, sys
 
 
-def create_XY(config, conflict_gdf, polygon_gdf):
+def create_XY(config, polygon_gdf, conflict_gdf):
     """Top-level function to create the X-array and Y-array.
     If the XY-data was pre-computed and specified in cfg-file, the data is loaded.
     If not, variable values and conflict data are read from file and stored in array. The resulting array is by default saved as npy-format to file.
@@ -26,18 +26,26 @@ def create_XY(config, conflict_gdf, polygon_gdf):
         XY = data.fill_XY(XY, config, conflict_gdf, polygon_gdf)
 
         if config.getboolean('general', 'verbose'): 
-            print('saving XY data by default to file {}'.format(os.path.abspath(os.path.join(config.get('general', 'input_dir'), 'XY.npy'))) + os.linesep)
+            print('INFO: saving XY data by default to file {}'.format(os.path.abspath(os.path.join(config.get('general', 'input_dir'), 'XY.npy'))) + os.linesep)
         np.save(os.path.join(config.get('general', 'input_dir'),'XY'), XY)
 
     else:
 
         if config.getboolean('general', 'verbose'): 
-            print('loading XY data from file {}'.format(os.path.abspath(os.path.join(config.get('general', 'input_dir'), config.get('pre_calc', 'XY')))) + os.linesep)
+            print('INFO: loading XY data from file {}'.format(os.path.abspath(os.path.join(config.get('general', 'input_dir'), config.get('pre_calc', 'XY')))) + os.linesep)
         XY = np.load(os.path.join(config.get('general', 'input_dir'), config.get('pre_calc', 'XY')), allow_pickle=True)
         
     X, Y = data.split_XY_data(XY, config)    
 
     return X, Y
+
+def create_X(config, polygon_gdf, conflict_gdf=None):
+
+    X = data.initiate_X_data(config)
+
+    X = data.fill_XY(X, config, conflict_gdf, polygon_gdf, make_proj=True)
+
+    return X
 
 def prepare_ML(config):
     """Top-level function to instantiate the scaler and model as specified in model configurations.
@@ -56,7 +64,7 @@ def prepare_ML(config):
 
     return scaler, clf
 
-def run(X, Y, config, scaler, clf, out_dir):
+def run_reference(X, Y, config, scaler, clf, out_dir):
     """Top-level function to run one of the four supported models.
 
     Args:
@@ -88,3 +96,12 @@ def run(X, Y, config, scaler, clf, out_dir):
         raise ValueError('the specified model type in the cfg-file is invalid - specify either 1, 2, 3 or 4.')
 
     return X_df, y_df, eval_dict
+
+def run_prediction(X, scaler, clf, config):
+
+    if config.getint('general', 'model') != 1:
+        raise ValueError('ERROR: making a prediction is only possible with model type 1, i.e. using all data')
+
+    y_df = models.predictive(X, scaler, clf, config)
+
+    return y_df
