+++ example/example_settings.cfg
@@ -13,6 +13,12 @@ y_start=2000
 y_end=2015
 # number of repetitions
 n_runs=50
+# whether to project
+make_proj=True
+# start year
+y_start_pred=2000
+# end year
+y_end_pred=2015
 
 [pre_calc]
 # if nothing is specified, the XY array will be stored in input_dir
@@ -35,7 +41,14 @@ shp=KoeppenGeiger/2000/1976-2000.shp
 zones=BWh,BSh
 code2class=KoeppenGeiger/classification_codes.txt
 
-[env_vars]
+[reference_data]
+# variable name here needs to be identical with variable name in nc-file
+total_evaporation=totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc
+precipitation=precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
+temperature=temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
+irr_water_demand=irrWaterDemand.nc
+
+[projection_data]
 # variable name here needs to be identical with variable name in nc-file
 total_evaporation=totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc
 precipitation=precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
