### 2020-11-06 8808a5b
updated functions and added code to make predictions
