+++ conflict_model/plots.py
@@ -1,234 +1,0 @@
-import matplotlib.pyplot as plt
-import geopandas as gpd
-import seaborn as sbs
-import numpy as np
-import os, sys
-from sklearn import metrics
-from conflict_model import evaluation
-
-def plot_active_polys(conflict_gdf, polygon_gdf, config, out_dir, **kwargs):
-    """Creates a (1,2)-subplot showing a) selected conflicts and all polygons, and b) selected conflicts and selected polygons.
-
-    Args:
-        conflict_gdf (geo-dataframe): geo-dataframe containing the selected conflicts.
-        extent_gdf (geo-dataframe): geo-dataframe containing all polygons.
-        polygon_gdf (geo-dataframe): geo-dataframe containing the selected polygons.
-        config (ConfigParser-object): object containing the parsed configuration-settings of the model.
-        out_dir (str): path to output folder
-    """    
-
-    fig, ax = plt.subplots(1, 1, **kwargs)
-    fig.suptitle('conflict distribution; # conflicts {}; threshold casualties {}; type of violence {}'.format(len(conflict_gdf), config.get('conflict', 'min_nr_casualties'), config.get('conflict', 'type_of_violence')))
-
-    conflict_gdf.plot(ax=ax, c='r', column='best', cmap='magma', 
-                      vmin=int(config.get('conflict', 'min_nr_casualties')), vmax=conflict_gdf.best.mean(), 
-                      legend=True, 
-                      legend_kwds={'label': "# casualties",})
-    polygon_gdf.boundary.plot(ax=ax)
-                            
-    plt.savefig(os.path.join(out_dir, 'selected_conflicts_and_polygons.png'), dpi=300)
-
-    return
-
-def plot_metrics_distribution(out_dict, out_dir, **kwargs):
-    """Plots the value distribution of a range of evaluation metrics based on all model simulations.
-
-    Args:
-        out_dict (dict): dictionary containing metrics score for various metrics and all simulation.
-        out_dir (str): path to output folder.
-    """    
-
-    fig, ax = plt.subplots(1, 1, **kwargs)
-
-    sbs.distplot(out_dict['Accuracy'], ax=ax, color="k", label='Accuracy')
-    sbs.distplot(out_dict['Precision'], ax=ax, color="r", label='Precision')
-    sbs.distplot(out_dict['Recall'], ax=ax, color="b", label='Recall')
-    plt.legend()
-
-    plt.savefig(os.path.join(out_dir, 'metrics_distribution.png'), dpi=300)
-
-    return
-
-def plot_nr_and_dist_pred(df, gdf, polygon_gdf, out_dir, **kwargs):
-    """Plots the number of number of predictions made per unique polygon, and the overall value distribution.
-
-    Args:
-        df (dataframe): containing model evaluation per unique polygon.
-        gdf (geo-dataframe): containing model evaluation per unique polygon.
-        polygon_gdf (geo-dataframe): geo-dataframe containing the selected polygons.
-        out_dir (str): path to output folder.
-        suffix (str, optional): suffix that can be used to discriminate between different analyses. Defaults to ''.
-    """    
-
-    fig, (ax1, ax2) = plt.subplots(1, 2, **kwargs)
-
-    gdf.plot(ax=ax1, column='ID_count', legend=True, cmap='cool')
-    polygon_gdf.boundary.plot(ax=ax1, color='0.5')
-    ax1.set_title('number of predictions made per polygon')
-    sbs.distplot(df.ID_count.values, ax=ax2)
-    ax2.set_title('distribution of predictions')
-
[…]
