### 2020-11-06 16b2555
fixed travisCI build with py3.6 but py3.8 still having problems with rasterio
