+++ .travis.yml
@@ -2,7 +2,7 @@
 
 language: python
 python:
-  - 3.8
+  # - 3.8
   - 3.7
   - 3.6
 
