### 2020-11-06 bdc5dfb
updated code to first fit clf based on ref period and then use this for pred
