### 2020-11-06 0962980
updated code to first fit clf based on ref period and then use this for pred
