### 2020-11-06 190439b
updated notebooks
