+++ example/example_settings.cfg
@@ -16,7 +16,7 @@ n_runs=50
 # whether to project
 make_proj=True
 # start year
-y_start_pred=2000
+y_start_pred=2010
 # end year
 y_end_pred=2015
 
