### 2020-11-06 9067c5e
added extra cfg-file for projections
