### 2020-11-06 e5e7002
changed cfg-file to make it as applicable for both ref and proj runs
