+++ example/example_settings.cfg
@@ -5,6 +5,8 @@ output_dir=../example/OUT
 # Note that only 1 supports sensitivity_analysis
 model=1
 verbose=True
+# ref for reference run or proj for projection run
+type=ref
 
 [settings]
 # start year
@@ -13,17 +15,14 @@ y_start=2000
 y_end=2015
 # number of repetitions
 n_runs=50
-# whether to project
-make_proj=True
-# start year
-y_start_pred=2010
-# end year
-y_end_pred=2015
 
 [pre_calc]
-# if nothing is specified, the XY array will be stored in input_dir
+# if nothing is specified, the XY array will be stored in output_dir
 # if XY already pre-calculated, then provide path to npy-file
 XY=
+# if nothing is specified, the classifier will be stored in output_dir
+# if classifier is already stored, then provide path to pkl-file
+clf=
 
 [extent]
 shp=waterProvinces/waterProvinces_Africa.shp
@@ -41,14 +40,7 @@ shp=KoeppenGeiger/2000/1976-2000.shp
 zones=BWh,BSh
 code2class=KoeppenGeiger/classification_codes.txt
 
-[reference_data]
-# variable name here needs to be identical with variable name in nc-file
-total_evaporation=totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc
-precipitation=precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
-temperature=temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
-irr_water_demand=irrWaterDemand.nc
-
-[projection_data]
+[data]
 # variable name here needs to be identical with variable name in nc-file
 total_evaporation=totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc
 precipitation=precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
