### 2020-11-06 9535067
added code to pickle clf if needed and load it later
