### 2020-11-06 59b33f5
added code to pickle clf if needed and load it later
