### 2020-11-09 b0301d8
aligned command line script to changes in code
