+++ scripts/runner.py
@@ -7,8 +7,7 @@ import os, sys
 import matplotlib.pyplot as plt
 
 import warnings
-warnings.filterwarnings("module")
-
+warnings.filterwarnings("ignore")
 
 @click.group()
 def cli():
@@ -24,14 +23,10 @@ def main(cfg):
         CFG (str): (relative) path to cfg-file
     """    
 
-    print('')
-    print('#### CONFLICT MODEL version {} ####'.format(copro.__version__))
-    print('')
-
     #- parsing settings-file and getting path to output folder
     config, out_dir = copro.utils.initiate_setup(cfg)
 
-    print('verbose mode on: {}'.format(config.getboolean('general', 'verbose')) + os.linesep)
+    if config.getboolean('general', 'verbose'): warnings.filterwarnings("default")
 
     #- selecting conflicts and getting area-of-interest and aggregation level
     conflict_gdf, extent_gdf, extent_active_polys_gdf, global_df = copro.selection.select(config, out_dir)
@@ -43,7 +38,7 @@ def main(cfg):
 
     #- create X and Y arrays by reading conflict and variable files;
     #- or by loading a pre-computed array (npy-file)
-    X, Y = copro.pipeline.create_XY(config, conflict_gdf, extent_active_polys_gdf)
+    X, Y = copro.pipeline.create_XY(config, extent_active_polys_gdf, conflict_gdf)
 
     #- defining scaling and model algorithms
     scaler, clf = copro.pipeline.prepare_ML(config)
@@ -61,11 +56,10 @@ def main(cfg):
     #- go through all n model executions
     for n in range(config.getint('settings', 'n_runs')):
         
-        if config.getboolean('general', 'verbose'):
-            print('run {} of {}'.format(n+1, config.getint('settings', 'n_runs')) + os.linesep)
+        click.echo('INFO: run {} of {}'.format(n+1, config.getint('settings', 'n_runs')))
 
         #- run machine learning model and return outputs
-        X_df, y_df, eval_dict = copro.pipeline.run(X, Y, config, scaler, clf, out_dir)
+        X_df, y_df, eval_dict = copro.pipeline.run_reference(X, Y, config, scaler, clf, out_dir)
         
         #- append per model execution
         #TODO: put all this into one function
@@ -89,7 +83,7 @@ def main(cfg):
     #- print mean values of all evaluation metrics
     for key in out_dict:
         if config.getboolean('general', 'verbose'):
-            print('average {0} of run with {1} repetitions is {2:0.3f}'.format(key, config.getint('settings', 'n_runs'), np.mean(out_dict[key])))
+            click.echo('DEBUG: average {0} of run with {1} repetitions is {2:0.3f}'.format(key, config.getint('settings', 'n_runs'), np.mean(out_dict[key])))
 
     # create accuracy values per polygon and save to output folder
     df_hit, gdf_hit = copro.evaluation.polygon_model_accuracy(out_y_df, global_df, out_dir)
@@ -110,6 +104,10 @@ def main(cfg):
     fig, ax = plt.subplots(1, 1)
     copro.plots.polygon_categorization(gdf_hit, ax=ax)
     plt.savefig(os.path.join(out_dir, 'polygon_categorization.png'), dpi=300, bbox_inches='tight')
+
+    copro.machine_learning.pickle_clf(scaler, clf, config)
+
+    click.echo('INFO: model run succesfully finished')
 
 if __name__ == '__main__':
     main()
