### 2020-11-09 b9c50af
replaced print with echo.click and output to stderr
