### 2020-11-09 4cdbee7
added function to plot header at model start
