### 2020-11-09 07b29e8
smaller print updates
