### 2020-11-09 cd26a1e
added color to click.echo output and updated runner.py to make porjection
