+++ scripts/run_script.sh
@@ -1,1 +1,1 @@
-python runner.py ../example/example_settings.cfg
+python runner.py ../example/example_settings.cfg -proj ../example/example_settings_proj.cfg
