+++ copro/models.py
@@ -193,12 +193,12 @@ def predictive(X, scaler, config):
     ##- scaling only the variable values
     X_ft = scaler.fit_transform(X_data)
 
-    if os.path.isfile(config.get('pre_calc', 'clf')):
-        with open(config.get('pre_calc', 'clf'), 'rb') as f:
-            print('INFO: loading classifier from {}'.format(config.get('pre_calc', 'clf')))
+    if os.path.isfile(os.path.abspath(config.get('pre_calc', 'clf'))):
+        with open(os.path.abspath(config.get('pre_calc', 'clf')), 'rb') as f:
+            print('INFO: loading classifier from {}'.format(os.path.abspath(config.get('pre_calc', 'clf'))))
             clf = pickle.load(f)
     else:
-        raise ValueError('ERROR: no pre-computed classifier specified in cfg-file')
+        raise ValueError('ERROR: no pre-computed classifier specified in cfg-file, currently specified file {} does not exist'.format(os.path.abspath(config.get('pre_calc', 'clf'))))
         
     print('INFO: making the projection')
     y_pred = clf.predict(X_ft)
