+++ example/example_settings_proj.cfg
@@ -1,6 +1,6 @@
 [general]
 input_dir=../example/example_data
-output_dir=../example/OUT
+output_dir=../example/OUT_PROJ
 # 1: all data; 2: leave-one-out model; 3: single variable model; 4: dubbelsteenmodel
 # Note that only 1 supports sensitivity_analysis
 model=1
@@ -17,10 +17,10 @@ n_runs=50
 [pre_calc]
 # if nothing is specified, the XY array will be stored in output_dir
 # if XY already pre-calculated, then provide (absolute) path to npy-file
-XY=../OUT/XY.npy
+XY=
 # if nothing is specified, the classifier will be stored in output_dir
 # if classifier is already stored, then provide (absolute) path to pkl-file
-clf=../OUT/clf.pkl
+clf=../example/OUT/clf.pkl
 
 [extent]
 shp=waterProvinces/waterProvinces_Africa.shp
