+++ example/example_settings.cfg
@@ -14,7 +14,7 @@ y_start=2000
 # end year
 y_end=2015
 # number of repetitions
-n_runs=50
+n_runs=3
 
 [pre_calc]
 # if nothing is specified, the XY array will be stored in output_dir
