### 2020-11-09 5c5ae4d
added color to click.echo output and updated runner.py to make porjection
