+++ copro/__init__.py
@@ -11,6 +11,6 @@ from . import evaluation
 from . import models
 from . import plots
 
-__author__ = """Jannis M. Hoch, Niko Wanders, Sophie de Bruin"""
+__author__ = """Jannis M. Hoch, Sophie de Bruin, Niko Wanders"""
 __email__ = 'j.m.hoch@uu.nl'
 __version__ = '0.0.6b'
