+++ example/example_settings_proj.cfg
@@ -4,7 +4,7 @@ output_dir=../example/OUT_PROJ
 # 1: all data; 2: leave-one-out model; 3: single variable model; 4: dubbelsteenmodel
 # Note that only 1 supports sensitivity_analysis
 model=1
-verbose=True
+verbose=False
 
 [settings]
 # start year
