+++ example/example_settings.cfg
@@ -4,9 +4,7 @@ output_dir=../example/OUT
 # 1: all data; 2: leave-one-out model; 3: single variable model; 4: dubbelsteenmodel
 # Note that only 1 supports sensitivity_analysis
 model=1
-verbose=True
-# ref for reference run or proj for projection run
-type=ref
+verbose=False
 
 [settings]
 # start year
