+++ copro/pipeline.py
@@ -96,6 +96,19 @@ def run_reference(X, Y, config, scaler, clf, out_dir):
     return X_df, y_df, eval_dict
 
 def run_prediction(X, scaler, config):
+    """Top-level function to run a predictive model with a already fitted classifier and new data.
+
+    Args:
+        X (array): X-array containing variable values.
+        scaler (scaler): the specified scaler instance.
+        config (ConfigParser-object): object containing the parsed configuration-settings of the model.
+
+    Raises:
+        ValueError: raised if another model type than the one using all data is specified in cfg-file.
+
+    Returns:
+        datatrame: containing model output on polygon-basis.
+    """    
 
     if config.getint('general', 'model') != 1:
         raise ValueError('ERROR: making a prediction is only possible with model type 1, i.e. using all data')
