+++ docs/api/models.rst
@@ -11,7 +11,9 @@ The various models
    models.leave_one_out
    models.single_variables
    models.dubbelsteen
+   models.predictive
 
 .. note::
 
-    the 'leave_one_out' and 'single_variables' models are only tested in beta-state.
+    The 'leave_one_out' and 'single_variables' models are only tested in beta-state. 
+    They will most likely be deprecated in near future.
