### 2020-11-09 937b468
updated model description
