### 2020-11-09 78e2494
updated model description
