+++ docs/examples/index.rst
@@ -5,7 +5,7 @@ Workflow
 
 This page provides a short example workflow in Jupyter Notebooks. 
 It is designed such that the main features of model become clear.
-Even though the model can be perfectly executed using notebooks, there is also the possibility to use a command line script (see :ref:`clickscript`).
+Even though the model can be perfectly executed using notebooks, there is also the possibility to use a command line script (see :ref:`script`).
 
 .. toctree::
     :maxdepth: 1
