+++ docs/model_settings.rst
@@ -1,3 +1,5 @@
+.. _settings:
+
 Model settings
 =========================
 
