### 2020-11-09 #2730 @JannisHoch
@sbenthall
thanks for not giving up the search for reviewers!
Not sure whom to contact, but I drafted an updated release of the code. It contains a few minor changes to the code and the JOSS manuscript is not affected by it.

Link to it: https://github.com/JannisHoch/copro/releases/tag/v0.0.6b.

Many thanks and kind regards, Jannis
