### 2020-11-13 56aa3dd
extende debugging and saving X array to file too when doing projections
