### 2020-11-13 2f6d884
extende debugging and saving X array to file too when doing projections
