### 2020-11-13 c67161d
improved naming of evaluation columns
