### 2020-11-13 e2e0da9
removed unncessary writing to out.txt
