### 2020-11-13 #2730 @sbenthall
@JannisHoch I believe that updates to the software are encouraged. It's possible that the review will motivate further updates.

This package is actually very much in line with my own research.

@kthyng is it possible for editors to also be reviewers on a piece, or is that a conflict of interest?
