+++ copro/utils.py
@@ -141,7 +141,7 @@ def print_model_info():
     click.echo('')
     click.echo(click.style('#### CoPro version {} ####'.format(copro.__version__), fg='yellow'))
     click.echo(click.style('#### For information about the model, please visit https://copro.readthedocs.io/ ####', fg='yellow'))
-    click.echo(click.style('#### Copyright ({}): {} ####'.format(date.today().year, copro.__author__), fg='yellow'))
+    click.echo(click.style('#### Copyright (2020-{}): {} ####'.format(date.today().year, copro.__author__), fg='yellow'))
     click.echo(click.style('#### Contact via: {} ####'.format(copro.__email__), fg='yellow'))
     click.echo(click.style('#### The model can be used and shared under the MIT license ####' + os.linesep, fg='yellow'))
 
