### 2020-11-13 #2730 @kthyng
> @kthyng is it possible for editors to also be reviewers on a piece, or is that a conflict of interest?

Yes that can happen though usually it is done because a reviewer dropped out or similar. If you're really interested in the package, I think it's fine for you to review if you'd like to.
