### 2020-11-17 #2730 @soodoku
so @kthyng, need some help. there used to be a checklist that i got to tick off. i can't see it here. have the procedures changed?
