### 2020-11-17 #2730 @soodoku
Thanks! Anything for me to do for now?

On Tue, Nov 17, 2020, 8:02 AM Kristen Thyng <notifications@github.com>
wrote:

> @soodoku <https://github.com/soodoku> No you’re correct about the
> checklist. However this is still the pre review not the review issue.
> @sbenthall <https://github.com/sbenthall> will start the actual review
> issue probably once two reviewers have been secured.
>
