### 2020-11-17 e1a3d38
split up scripts for reference and reference+projections
