### 2020-11-17 a5f0577
split up scripts for reference and reference+projections
