+++ scripts/run_reference.sh
@@ -1,0 +1,1 @@
+python runner.py ../example/example_settings.cfg
