### 2020-11-17 3e3e939
outputting trps and aucs for ROC plot to csv
