+++ CITATION.cff
@@ -1,23 +1,35 @@
 # YAML 1.2
-# Metadata for citation of this software according to the CFF format (https://citation-file-format.github.io/)
-cff-version: 1.0.3
-message: If you use this software, please cite it using these metadata.
+---
+cff-version: "1.1.0"
+message: "If you use this software, please cite it using these metadata."
 title: CoPro
-doi: 10.5281/zenodo.4264684
-authors:
-- given-names: Jannis
-  family-names: Hoch
-  affiliation: Utrecht University
-  orcid: https://orcid.org/0000-0003-3570-6436
-- given-names: Sophie
-  family-names: de Bruin
-  affiliation: PBL Netherlands Environmental Assessment Agency
-  orcid: https://orcid.org/0000-0003-3429-349X
-- given-names: Niko
-  family-names: Wanders
-  affiliation: Utrecht University
-  orcid: https://orcid.org/0000-0002-7102-5454
-version: 0.0.6b
+version: "v0.0.6b"
+authors: 
+  -
+    affiliation: "Utrecht University"
+    family-names: Hoch
+    given-names: Jannis
+    orcid: "https://orcid.org/0000-0003-3570-6436"
+  -
+    affiliation: "PBL Netherlands Environmental Assessment Agency"
+    family-names: "de Bruine"
+    given-names: Sophie
+    orcid: "https://orcid.org/0000-0003-3429-349X"
+  -
+    affiliation: "Utrecht University"
+    family-names: Wanders
+    given-names: Niko
+    orcid: "https://orcid.org/0000-0002-7102-5454"
 date-released: 2020-11-09
-repository-code: https://github.com/JannisHoch/copro
+doi: "10.5281/zenodo.4264684"
 license: MIT
+repository-code: "https://github.com/JannisHoch/copro"
+keywords: 
+  - conflict
+  - "machine learning"
+  - "climate change"
+  - water
+  - scenarios
+  - projections
+  - "climate security"
+...
