### 2020-11-17 be01275
deleting all files in out_dir if already exists
