### 2020-11-17 ebf3eee
towards version 0.0.6b-joss
