+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.0.6b
+current_version = 0.0.6b-joss
 commit = True
 tag = True
 
