### 2020-11-17 cc865b6
towards version 0.0.6b-joss
