+++ README.rst
@@ -67,12 +67,15 @@ Runner script
 To run the model from command line, a command line script is provided. 
 All data and settings are retrieved from the settings-file which needs to be provided as inline argument.
 
+There are two settings-files, one for evaluating the model for the reference situation, and another one for additionally making projections.
+
 .. code-block:: console
 
     $ cd path/to/copro/scripts
     $ python runner.py ../example/example_settings.cfg
+    $ python runner.py ../example/example_settings_proj.cfg
 
-By default, output is stored to the output directory specified in the settings-file. 
+By default, output is stored to the output directory specified in the specific settings-file. 
 
 Documentation
 ---------------
