### 2020-11-18 05f8ed8
Bump notebook from 6.0.3 to 6.1.5 in /docs
Bumps [notebook](https://github.com/jupyter/jupyterhub) from 6.0.3 to 6.1.5.
- [Release notes](https://github.com/jupyter/jupyterhub/releases)
- [Changelog](https://github.com/jupyterhub/jupyterhub/blob/master/CHECKLIST-Release.md)
- [Commits](https://github.com/jupyter/jupyterhub/commits)

Signed-off-by: dependabot[bot] <support@github.com>
