+++ docs/requirements.rtd.txt
@@ -18,7 +18,7 @@ rasterio==1.1.3
 rioxarray==0.0.26
 xarray==0.15.0
 ipython==7.13.0
-notebook==6.0.3
+notebook==6.1.5
 nbconvert==5.6.1
 netCDF4==1.5.3
 scikit-learn==0.22.1
