### 2020-11-19 f0a0159
bugfixes, eg plotting for linux envs
