+++ copro/plots.py
@@ -1,3 +1,5 @@
+import matplotlib
+matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 import geopandas as gpd
 import seaborn as sbs
