### 2020-11-19 ba8809a
bugfixes, eg plotting for linux envs
