+++ copro/conflict.py
@@ -1,7 +1,6 @@
 import geopandas as gpd
 import pandas as pd
 import numpy as np
-import matplotlib.pyplot as plt
 import os, sys
 
 def conflict_in_year_bool(conflict_gdf, extent_gdf, config, sim_year): 
