+++ copro/evaluation.py
@@ -4,7 +4,6 @@ from sklearn import metrics
 import pandas as pd
 import geopandas as gpd
 import numpy as np
-import matplotlib.pyplot as plt
 
 def init_out_dict():
     """Initiates the main model evaluatoin dictionary for a range of model metric scores. 
