+++ setup.py
@@ -45,6 +45,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/JannisHoch/copro',
-    version='0.0.6b-joss',
+    version='0.0.6',
     zip_safe=False,
 )
