+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.0.6b-joss
+current_version = 0.0.6
 commit = True
 tag = True
 
