+++ README.rst
@@ -17,7 +17,7 @@ Welcome to CoPro, a machine-learning tool for conflict risk projections based on
     :target: https://copro.readthedocs.io/en/latest/?badge=latest
 
 .. image:: https://img.shields.io/github/v/release/JannisHoch/copro
-    :target: https://github.com/JannisHoch/copro/releases/tag/v0.0.5-pre
+    :target: https://github.com/JannisHoch/copro/releases/tag/v0.0.6
 
 .. image:: https://zenodo.org/badge/254407279.svg
     :target: https://zenodo.org/badge/latestdoi/254407279
