### 2020-11-27 cdaf489
towards v0.0.7 which should be the final version after joss review
