### 2020-11-27 c8379dc
renamed command line script
