### 2020-11-27 545f330
finalizing readme.md for now
