### 2020-11-27 9c702d7
added model purpose
