### 2020-11-30 #2855 @JannisHoch
HI @soodoku and @sbenthall, in the meantime, the model evolved to version 0.0.6 (https://github.com/JannisHoch/copro/releases/tag/v0.0.6). All changes made in the process of this review will be committed to the branch 'joss-review' (https://github.com/JannisHoch/copro/tree/joss_review).
