+++ example/example_settings_proj.cfg
@@ -40,10 +40,10 @@ code2class=KoeppenGeiger/classification_codes.txt
 
 [data]
 # variable name here needs to be identical with variable name in nc-file
-total_evaporation=totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc
-precipitation=precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
-temperature=temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
-irr_water_demand=irrWaterDemand.nc
+total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc
+precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
+temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
+irr_water_demand=hydro/irrWaterDemand.nc
 
 [machine_learning]
 # choose from: MinMaxScaler, StandardScaler, RobustScaler, QuantileTransformer
