### 2020-11-30 cad1692
updated cfg-files based on data moving
