+++ example/example_settings.cfg
@@ -33,17 +33,17 @@ min_nr_casualties=1
 type_of_violence=1,2,3
 
 [climate]
-shp=KoeppenGeiger/2000/1976-2000.shp
+shp=KoeppenGeiger/2000/Koeppen_Geiger_1976-2000.shp
 # define either one or more classes (use abbreviations!) or specify None for not filtering
 zones=BWh,BSh
 code2class=KoeppenGeiger/classification_codes.txt
 
 [data]
 # variable name here needs to be identical with variable name in nc-file
-total_evaporation=totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc
-precipitation=precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
-temperature=temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
-irr_water_demand=irrWaterDemand.nc
+total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc
+precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
+temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
+irr_water_demand=hydro/irrWaterDemand.nc
 
 [machine_learning]
 # choose from: MinMaxScaler, StandardScaler, RobustScaler, QuantileTransformer
