### 2020-11-30 20405e7
updated run.sh to remove tmp files after running
