### 2020-11-30 36bb1bb
all bash-files now in example folder
