+++ example/run_script_projections.sh
@@ -1,0 +1,1 @@
+python ../scripts/copro_runner.py example_settings.cfg -proj example_settings_proj.cfg
