+++ example/example_settings.cfg
@@ -1,6 +1,6 @@
 [general]
-input_dir=../example/example_data
-output_dir=../example/OUT
+input_dir=./example_data
+output_dir=./OUT
 # 1: all data; 2: leave-one-out model; 3: single variable model; 4: dubbelsteenmodel
 # Note that only 1 supports sensitivity_analysis
 model=1
@@ -27,7 +27,7 @@ shp=waterProvinces/waterProvinces_Africa.shp
 
 [conflict]
 # either specify path to file or state 'download' to download latest PRIO/UCDP dataset
-conflict_file=download
+conflict_file=UCDP/ged201.csv
 min_nr_casualties=1
 # 1=state-based armed conflict; 2=non-state conflict; 3=one-sided violence
 type_of_violence=1,2,3
