+++ scripts/run_reference.sh
@@ -1,1 +1,0 @@
-python copro_runner.py ../example/example_settings.cfg
