+++ scripts/run_projections.sh
@@ -1,1 +1,0 @@
-python copro_runner.py ../example/example_settings.cfg -proj ../example/example_settings_proj.cfg
