### 2020-11-30 919f396
updated doc with pip install -e instructions
