### 2020-11-30 a1afa2b
updated doc with pip install -e instructions
