+++ docs/Installation.rst
@@ -1,21 +1,34 @@
 Installation
 =========================
 
-From GitHub
--------------
-
-To install the conflict model, first clone the code from GitHub. It is advised to create an individual python environment first. 
-You can then install the model package into this environment.
+To install copro from GitHub, first clone the code. It is advised to create an individual python environment first. 
 
 .. code-block:: console
 
-    $ git clone https://github.com/JannisHoch/copro.git
-    $ cd path/to/copro
-    $ conda env create -f environment.yml
-    $ conda activate copro
-    $ python setup.py develop
+    git clone https://github.com/JannisHoch/copro.git
+    cd path/to/copro
+    conda env create -f environment.yml
+
+It is now possible to activate this environment with
+
+.. code-block:: console
+
+    conda activate copro
+
+To install CoPro in editable mode in this environment, run this command next
+
+.. code-block:: console
+
+    pip install -e <path/to/copro>/copro
 
 From PyPI
+------------
+
+.. todo::
+
+    This is not yet supported.
+
+From conda
 ------------
 
 .. todo::
