### 2020-11-30 c5ee67a
udpated model exeuction commands
