+++ joss/paper.md
@@ -29,14 +29,14 @@ bibliography: bibliography.bib
 
 # Summary
 
-Climate change and environmental degradation are increasingly recognised as factors that can contribute to conflict risk under specific conditions.
-In light of predicted shifts in climate patterns and the potentially resulting battle for increasingly scarce resources, it is widely acknowledged that there is an actual risk of increased armed conflict. To efficiently plan and implement adaptation and mitigation measures, it is key to first obtain an understanding of the the impact of individual climate, environmental, and societal drivers on conflict risk. And second, conflict risk needs to be projected to a given point in the future to be able to prepare accordingly. With CoPro, a set of functions and a workflow is made openly accessible to perform these steps, yielding maps of relative conflict risk. The model caters for a variety of settings and input data, thereby capturing the multitude of facets of the climate-environment-conflict nexus.
+Climate change and environmental degradation are increasingly recognized as factors that can contribute to conflict risk under specific conditions.
+In light of predicted shifts in climate patterns and the potentially resulting battle for increasingly scarce resources, it is widely acknowledged that there is an actual risk of increased armed conflict. To efficiently plan and implement adaptation and mitigation measures, it is key to first obtain an understanding of the impact of individual climate, environmental, and societal drivers on conflict risk. And second, conflict risk needs to be projected to a given point in the future to be able to prepare accordingly. With CoPro, a set of functions and a workflow is made openly accessible to perform these steps, yielding maps of relative conflict risk. The model caters for a variety of settings and input data, thereby capturing the multitude of facets of the climate-environment-conflict nexus.
 
 # Statement of need 
 
-There is increasing consensus that climate change can exacerbate the risk of (armed) conflict [@koubi2019climate; @mach2019climate]. Nevertheless, making (operational) forecasts on the short-term is still challenging due to several reasons [@cederman2017predicting]. Building upon recent, similar approaches to use data-driven models [@colaresi2017robot] and statistical approaches [@witmer2017subnational; @hegre2016forecasting] for conflict risk projections, CoPro is a novel, fully open, and extensible tool to combine the inter-disciplinary expertise required to make long-term projections of conflict risk associated with climatic and environmental drivers. Such a tool is needed not only to integrate the different disciplines, but also to extend the modelling approach with new insights and data - after all, the established links between climate and societal factors with conflict are still weak [@koubi2019climate; @mach2019climate]. In addition to scholarly explorations of the inter-dependencies and importance of various conflict drivers, model output such as maps of spatially-disaggregated projected conflict risk can be an invaluable input to inform the decision-making process in affected regions.
+There is increasing consensus that climate change can exacerbate the risk of (armed) conflict [@koubi2019climate; @mach2019climate]. Nevertheless, making (operational) forecasts on the short-term is still challenging due to several reasons [@cederman2017predicting]. Building upon recent, similar approaches to use data-driven models [@colaresi2017robot] and statistical approaches [@witmer2017subnational; @hegre2016forecasting] for conflict risk projections, CoPro is a novel, fully open, and extensible tool to combine the inter-disciplinary expertise required to make long-term projections of conflict risk associated with climatic and environmental drivers. Such a tool is needed not only to integrate the different disciplines, but also to extend the modeling approach with new insights and data - after all, the established links between climate and societal factors with conflict are still weak [@koubi2019climate; @mach2019climate]. In addition to scholarly explorations of the inter-dependencies and importance of various conflict drivers, model output such as maps of spatially-disaggregated projected conflict risk can be an invaluable input to inform the decision-making process in affected regions.
 
[…]
