+++ docs/model_settings.rst
@@ -3,12 +3,8 @@
 Model settings
 =========================
 
-The main model settings need to be specified in a configuration file. Here, the different sections are explained briefly. 
-Additionally, the cfg-file included in the GitHub-repository already contains the key information.
-
-.. note::
-
-    The file-type (we here refer to the cfg-file) is less relevant as long as the file can be parsed with ConfigParser.
+The main model settings need to be specified in a configuration file (cfg-file). 
+Here, the different sections are explained briefly. 
 
 **[general]**
 
@@ -25,17 +21,18 @@ Additionally, the cfg-file included in the GitHub-repository already contains th
 
     the 'leave-one-out' and 'single variables' models are only tested in beta-state. They produce only limited output (see :ref:`Output`). 
 
-- *verbose*: if True, model messages will be printed to screen.
+- *verbose*: if True, additional messages will be printed.
 
 **[settings]**
 
 - *y_start*: the start year of the simulation;
 - *y_end*: the end year of the simulation. All data between y_start and y_end will be used to train and test the model;
-- *n_runs*: the number of model executions. By repeating the split, train, test steps multiple times, coincidental results can be avoided.
+- *n_runs*: the number repetitions of the split-sample test for training and testing the model. By repeating these steps multiple times, coincidental results can be avoided.
 
 **[pre_calc]**
 
-- *XY*: if the XY-data was already pre-computed in a previous run and stored as npy-file, it be specified here and will be loaded from file. If nothing is specified, the model will save the XY-data by default to the input directory.
+- *XY*: if the XY-data was already pre-computed in a previous run and stored as npy-file, it can be specified here and will be loaded from file. If nothing is specified, the model will save the XY-data by default to the output directory as ``XY.npy``;
+- *clf*: path to the pickled fitted classifier from the reference run. Needed for projection runs only;
 
 **[extent]**
 
@@ -55,9 +52,9 @@ Additionally, the cfg-file included in the GitHub-repository already contains th
     2. non-state conflict: the use of armed force between two organized armed groups, neither of which is the government of a state, which results in at least 25 battle-related deaths in a year;
     3. one-sided violence: the deliberate use of armed force by the government of a state or by a formally organized group against civilians which results in at least 25 deaths in a year.
 
-.. note::
+.. important::
 
-    Currently, the model works only with data from the `PRIO/UCDP programme <https://www.prio.org/Data/Armed-Conflict/UCDP-PRIO/>`_. Other conflict datasets such `ACLED <https://acleddata.com/>`_ are not supported due to differences in data categorization.
+    CoPro currently only works with UCDP data. As other data sources will be supported in the future, the conflict selection process will be come more elaborated.
 
 **[climate]**
 
@@ -67,7 +64,7 @@ Additionally, the cfg-file included in the GitHub-repository already contains th
 
 .. warning:: 
 
-    The code2class-file should not be altered except if another shp-file is provided containing undocumented climate classes.
+    The code2class-file should not be altered!
 
 **[reference_data]**
 
@@ -77,9 +74,4 @@ In this section, all variables to be used in the model need to be provided. The 
 
 - *scaler*: the scaling algorithm used to scale the variable values to comparable scales. Currently supported are ``MinMaxScaler``, ``StandardScaler``, ``RobustScaler``, and ``QuantileTransformer``;
 - *model*: the machine learning algorithm to be applied. Currently supported are ``NuSVC``, ``KNeighborsClassifier``, and ``RFClassifier``;
-
-.. note:: 
-
-    The parameters of the models were already calibrated using the sklearn `GridSearchCV <https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.GridSearchCV.html>`_.
-
[…]
