### 2020-11-30 07534a5
updated output documentation and some bugfixes
