+++ docs/examples/index.rst
@@ -4,8 +4,9 @@ Workflow
 =========
 
 This page provides a short example workflow in Jupyter Notebooks. 
-It is designed such that the main features of model become clear.
-Even though the model can be perfectly executed using notebooks, there is also the possibility to use a command line script (see :ref:`script`).
+It is designed such that the main features of CoPro become clear.
+
+Even though the model can be perfectly executed using notebooks, the main (and more convenient) way of model execution is the command line script (see :ref:`script`).
 
 .. toctree::
     :maxdepth: 1
