+++ docs/Output.rst
@@ -1,101 +1,56 @@
 Output
 =========================
 
-The model can produce a range of output files. Output is stored in the output folder as specified in the configurations-file (cfg-file).
+The model can produce a range of output files. All output is stored in the output folder as specified in the configurations-file (cfg-file).
 
-.. note:: 
-
-    In addition to these output files, the model settings file (cfg-file) is automatically copied to the output folder.
+In addition to the output files listed below, the model settings file (cfg-file) is automatically copied to the output folder.
 
 .. important:: 
 
     Not all model types provide the output mentioned below. If the 'leave-one-out' or 'single variable' model are selected, only the metrics are stored to a csv-file.
 
-.. important::
 
-    Most of the output can only be produced when running a reference model, i.e. when comparing the predictions against observations. 
-    If running a prediction model, only the chance of conflict per polygon is stored to file.
++-------------------------------+---------------------------------------------------------------------------------------------+---------------------------------------------------------------------------------------------+
+| File name                     | Description                                                                                 | Note                                                                                        |
++===============================+=============================================================================================+=============================================================================================+
+| ``selected_polygons.shp``     | Shapefile containing all remaining polygons after selection procedure                       |                                                                                             |
++-------------------------------+---------------------------------------------------------------------------------------------+---------------------------------------------------------------------------------------------+
+| ``selected_conflicts.shp``    | Shapefile containing all remaining conflict points after selection procedure                |                                                                                             | 
++-------------------------------+---------------------------------------------------------------------------------------------+---------------------------------------------------------------------------------------------+
+| ``XY.npy``                    | NumPy-array containing geometry, ID, and scaled data of sample (X) and target data (Y)      | can be provided in cfg-file to safe time in next run; file can be loaded with numpy.load()  | 
++-------------------------------+---------------------------------------------------------------------------------------------+---------------------------------------------------------------------------------------------+
+| ``X.npy``                     | NumPy-array containing geometry, ID, and scaled data of sample (X)                          | only written in projection run; file can be loaded with numpy.load()                        | 
++-------------------------------+---------------------------------------------------------------------------------------------+---------------------------------------------------------------------------------------------+
+| ``clf.pkl``                   | Pickled classifier fitted with the entirety of XY-data                                      | needed to perform projection run; file can be loaded with pickle.load()                     | 
++-------------------------------+---------------------------------------------------------------------------------------------+---------------------------------------------------------------------------------------------+
+| ``raw_output_data.npy``       | NumPy-array containing each single prediction made in the reference run                     | will contain multiple predictions per polygon; file can be loaded with numpy.load()         | 
++-------------------------------+---------------------------------------------------------------------------------------------+---------------------------------------------------------------------------------------------+
+| ``evaluation_metrics.csv``    | Various evaluation metrics determined per repetition of the split-sample test repetition    | file can e.g. be loaded with pandas.read_csv()                                              | 
++-------------------------------+---------------------------------------------------------------------------------------------+---------------------------------------------------------------------------------------------+
[…]
