### 2020-11-30 a9fce08
updated output documentation and some bugfixes
