### 2020-11-30 c458f92
added more contextual info and explanations to notebooks
