+++ docs/examples/index.rst
@@ -4,7 +4,9 @@ Workflow
 =========
 
 This page provides a short example workflow in Jupyter Notebooks. 
-It is designed such that the main features of CoPro become clear.
+It is designed such that the main steps, features, assumptions, and outcomes of CoPro become clear.
+
+As model input data, the data set downloadable from `Zenodo <https://zenodo.org/record/4297295>`_ was used. 
 
 Even though the model can be perfectly executed using notebooks, the main (and more convenient) way of model execution is the command line script (see :ref:`script`).
 
