### 2020-11-30 4fb1943
added more contextual info and explanations to notebooks
