+++ example/example_settings_proj.cfg
@@ -33,7 +33,7 @@ min_nr_casualties=1
 type_of_violence=1,2,3
 
 [climate]
-shp=KoeppenGeiger/2000/1976-2000.shp
+shp=KoeppenGeiger/2000/Koeppen_Geiger_1976-2000.shp
 # define either one or more classes (use abbreviations!) or specify None for not filtering
 zones=BWh,BSh
 code2class=KoeppenGeiger/classification_codes.txt
