### 2020-12-02 a31963e
clearer description of CoPro added value
