+++ environment.yml
@@ -25,6 +25,7 @@ dependencies:
   - seaborn==0.11.0
   - pip
   - colorama
+  - setuptools>=49.6
   - pip:
     - rasterstats==0.14.0
     
