### 2020-12-02 b1f8861
moved scripts into copro folder for executable
