### 2020-12-02 f4171ac
working with location of cfg-file as root dir now
