+++ copro/scripts/copro_runner.py
@@ -9,10 +9,6 @@ import matplotlib.pyplot as plt
 import warnings
 warnings.filterwarnings("ignore")
 
-# @click.group()
-# def cli():
-#     pass
-
 @click.command()
 @click.argument('cfg', type=click.Path())
 @click.option('--projection-settings', '-proj', help='path to cfg-file with settings for a projection run', multiple=True, type=click.Path())
@@ -29,7 +25,7 @@ def cli(cfg, projection_settings=[], verbose=False):
     """ 
   
     #- parsing settings-file and getting path to output folder
-    config, out_dir = copro.utils.initiate_setup(cfg)
+    config, out_dir, root_dir = copro.utils.initiate_setup(cfg)
 
     if verbose:
         config.set('general', 'verbose', str(verbose))
@@ -37,7 +33,7 @@ def cli(cfg, projection_settings=[], verbose=False):
     click.echo(click.style('\nINFO: reference run started\n', fg='cyan'))
 
     #- selecting conflicts and getting area-of-interest and aggregation level
-    conflict_gdf, extent_gdf, extent_active_polys_gdf, global_df = copro.selection.select(config, out_dir)
+    conflict_gdf, extent_gdf, extent_active_polys_gdf, global_df = copro.selection.select(config, out_dir, root_dir)
     #- plot selected polygons and conflicts
     fig, ax = plt.subplots(1, 1)
     copro.plots.selected_polygons(extent_active_polys_gdf, figsize=(20, 10), ax=ax)
@@ -46,7 +42,7 @@ def cli(cfg, projection_settings=[], verbose=False):
 
     #- create X and Y arrays by reading conflict and variable files;
     #- or by loading a pre-computed array (npy-file)
-    X, Y = copro.pipeline.create_XY(config, extent_active_polys_gdf, conflict_gdf)
+    X, Y = copro.pipeline.create_XY(config, out_dir, root_dir, extent_active_polys_gdf, conflict_gdf)
 
     #- defining scaling and model algorithms
     scaler, clf = copro.pipeline.prepare_ML(config)
@@ -104,7 +100,7 @@ def cli(cfg, projection_settings=[], verbose=False):
     copro.plots.metrics_distribution(out_dict, figsize=(20, 10))
     plt.savefig(os.path.join(out_dir, 'metrics_distribution.png'), dpi=300, bbox_inches='tight')
 
-    clf = copro.machine_learning.pickle_clf(scaler, clf, config)
+    clf = copro.machine_learning.pickle_clf(scaler, clf, config, root_dir)
     #- plot relative importance of each feature based on ALL data points
     fig, ax = plt.subplots(1, 1)
     copro.plots.factor_importance(clf, config, out_dir=out_dir, ax=ax, figsize=(20, 10))
@@ -118,10 +114,10 @@ def cli(cfg, projection_settings=[], verbose=False):
 
             click.echo(click.style('\nINFO: projection run started, based on {}'.format(os.path.abspath(proj)), fg='cyan'))
 
-            config, out_dir = copro.utils.initiate_setup(proj)
+            config, out_dir, root_dir = copro.utils.initiate_setup(proj)
 
-            X = copro.pipeline.create_X(config, extent_active_polys_gdf)
+            X = copro.pipeline.create_X(config, out_dir, root_dir, extent_active_polys_gdf)
 
-            y_df = copro.pipeline.run_prediction(X, scaler, config)
+            y_df = copro.pipeline.run_prediction(X, scaler, config, root_dir)
 
             df_hit, gdf_hit = copro.evaluation.polygon_model_accuracy(y_df, global_df, out_dir=out_dir, make_proj=True)
