### 2020-12-02 365d624
working with location of cfg-file as root dir now
