+++ copro/pipeline.py
@@ -4,13 +4,15 @@ import numpy as np
 import os, sys
 
 
-def create_XY(config, polygon_gdf, conflict_gdf):
+def create_XY(config, out_dir, root_dir, polygon_gdf, conflict_gdf):
     """Top-level function to create the X-array and Y-array.
     If the XY-data was pre-computed and specified in cfg-file, the data is loaded.
     If not, variable values and conflict data are read from file and stored in array. The resulting array is by default saved as npy-format to file.
 
     Args:
         config (ConfigParser-object): object containing the parsed configuration-settings of the model.
+        out_dir (str): path to output folder.
+        root_dir (str): path to location of cfg-file.
         conflict_gdf (geo-dataframe): geo-dataframe containing the selected conflicts.
         polygon_gdf (geo-dataframe): geo-dataframe containing the selected polygons.
 
@@ -23,35 +25,47 @@ def create_XY(config, polygon_gdf, conflict_gdf):
 
         XY = data.initiate_XY_data(config)
 
-        XY = data.fill_XY(XY, config, conflict_gdf, polygon_gdf)
+        XY = data.fill_XY(XY, config, root_dir, conflict_gdf, polygon_gdf)
 
-        print('INFO: saving XY data by default to file {}'.format(os.path.abspath(os.path.join(config.get('general', 'output_dir'), 'XY.npy'))))
-        np.save(os.path.join(config.get('general', 'output_dir'),'XY'), XY)
+        print('INFO: saving XY data by default to file {}'.format(os.path.join(out_dir, 'XY.npy')))
+        np.save(os.path.join(out_dir,'XY'), XY)
 
     else:
 
-        print('INFO: loading XY data from file {}'.format(os.path.abspath(os.path.join(config.get('general', 'output_dir'), config.get('pre_calc', 'XY')))))
-        XY = np.load(os.path.join(config.get('general', 'output_dir'), config.get('pre_calc', 'XY')), allow_pickle=True)
+        print('INFO: loading XY data from file {}'.format(os.path.join(root_dir, config.get('pre_calc', 'XY'))))
+        XY = np.load(os.path.join(root_dir, config.get('pre_calc', 'XY')), allow_pickle=True)
         
     X, Y = data.split_XY_data(XY, config)    
 
     return X, Y
 
-def create_X(config, polygon_gdf, conflict_gdf=None):
+def create_X(config, out_dir, root_dir, polygon_gdf, conflict_gdf=None):
+    """[summary]
+
+    Args:
+        config ([type]): [description]
+        out_dir ([type]): [description]
+        root_dir ([type]): [description]
+        polygon_gdf ([type]): [description]
+        conflict_gdf ([type], optional): [description]. Defaults to None.
+
+    Returns:
+        [type]: [description]
+    """    
 
     if config.get('pre_calc', 'XY') is '':
 
         X = data.initiate_X_data(config)
 
-        X = data.fill_XY(X, config, conflict_gdf, polygon_gdf)
+        X = data.fill_XY(X, config, root_dir, conflict_gdf, polygon_gdf)
 
-        print('INFO: saving X data by default to file {}'.format(os.path.abspath(os.path.join(config.get('general', 'output_dir'), 'X.npy'))))
-        np.save(os.path.join(config.get('general', 'output_dir'),'X'), X)
+        print('INFO: saving X data by default to file {}'.format(os.path.join(out_dir, 'X.npy')))
+        np.save(os.path.join(out_dir,'X'), X)
 
     else:
 
-        print('INFO: loading XY data from file {}'.format(os.path.abspath(config.get('pre_calc', 'X'))))
-        X = np.load(os.path.abspath(config.get('pre_calc', 'X')), allow_pickle=True)
[…]
