+++ example/example_settings_proj.cfg
@@ -20,7 +20,7 @@ n_runs=50
 XY=
 # if nothing is specified, the classifier will be stored in output_dir
 # if classifier is already stored, then provide (absolute) path to pkl-file
-clf=../example/OUT/clf.pkl
+clf=./OUT/clf.pkl
 
 [extent]
 shp=waterProvinces/waterProvinces_Africa.shp
