### 2020-12-02 cab004c
added docstring
