### 2020-12-02 3ec2b14
updates sh-script to exe
