### 2020-12-02 ef661f5
update installation and execution instructions
