+++ appveyor.yml
@@ -1,0 +1,9 @@
+# version format
+version: 1.0.{build}
+
+# branches to build
+branches:
+  # whitelist
+  only:
+    - master
+    - dev
