+++ appveyor.yml
@@ -1,10 +1,0 @@
-# version format
-version: 1.0.{build}
-
-# branches to build
-branches:
-  # whitelist
-  only:
-    - master
-    - dev
-    - joss_review
