### 2020-12-04 423ddaa
added random_state=42 to model and scaler
