### 2020-12-09 a07acaf
implemented conflict at t-1 as new sample and tested it
