### 2020-12-09 258a72b
implemented conflict at t-1 as new sample and tested it
