### 2020-12-11 47aebad
ln of a variable can be computed
