+++ copro/variables.py
@@ -5,6 +5,7 @@ import geopandas as gpd
 import rasterstats as rstats
 import numpy as np
 import os, sys
+import math
 
 import warnings
 warnings.filterwarnings("ignore")
@@ -37,19 +38,31 @@ def nc_with_float_timestamp(extent_gdf, config, root_dir, var_name, sim_year, st
     Returns:
         list: list containing statistical value per polygon, i.e. with same length as extent_gdf
     """   
-    # get path to netCDF-file.
-    # nc_fo = os.path.join(os.path.abspath(config.get('general', 'input_dir')), 
-    #                      config.get('data', var_name))
 
-    nc_fo = os.path.join(root_dir, config.get('general', 'input_dir'), config.get('data', var_name))
+    data_fo = os.path.join(root_dir, config.get('general', 'input_dir'), config.get('data', var_name)).rsplit(',')
+
+    if len(data_fo) == 1:
+        nc_fo = data_fo[0]
+        ln_flag = False
+    elif len(data_fo) == 2:
+        nc_fo = data_fo[0]
+        if data_fo[1] == 'ln':
+            ln_flag = True
+        else:
+            raise ValueError('ERROR: no valid ln flag set')
 
     if config.getboolean('general', 'verbose'): print('DEBUG: calculating mean {0} per aggregation unit from file {1} for year {2}'.format(var_name, nc_fo, sim_year))
+
+    print(nc_fo)
+    print(ln_flag)
 
     # open nc-file with xarray as dataset
     nc_ds = xr.open_dataset(nc_fo)
     # get xarray data-array for specified variable
     nc_var = nc_ds[var_name]
-
+    if ln_flag:
+        nc_var = np.log(nc_var)
+        print('DEBUG: log-transform variable {}'.format(var_name))
     # open nc-file with rasterio to get affine information
     affine = rio.open(nc_fo).transform
 
@@ -57,7 +70,7 @@ def nc_with_float_timestamp(extent_gdf, config, root_dir, var_name, sim_year, st
     nc_arr = nc_var.sel(time=sim_year)
     nc_arr_vals = nc_arr.values
     if nc_arr_vals.size == 0:
-        raise ValueError('the data was found for this year in the nc-file {}, check if all is correct'.format(nc_fo))
+        raise ValueError('ERROR: the data was found for this year in the nc-file {}, check if all is correct'.format(nc_fo))
 
     # initialize output list
     list_out = []
@@ -67,6 +80,9 @@ def nc_with_float_timestamp(extent_gdf, config, root_dir, var_name, sim_year, st
         zonal_stats = rstats.zonal_stats(prov.geometry, nc_arr_vals, affine=affine, stats=stat_func)
         if (zonal_stats[0][stat_func] == None) and (config.getboolean('general', 'verbose')): 
             print('WARNING: NaN computed!')
+
+        print(zonal_stats[0][stat_func])
+
         list_out.append(zonal_stats[0][stat_func])
 
     if config.getboolean('general', 'verbose'): print('DEBUG: ... done.')
@@ -99,22 +115,35 @@ def nc_with_continous_datetime_timestamp(extent_gdf, config, root_dir, var_name,
     Returns:
         list: list containing statistical value per polygon, i.e. with same length as extent_gdf
     """   
-    # get path to netCDF-file.
-    # nc_fo = os.path.join(os.path.abspath(config.get('general', 'input_dir')), 
-    #                      config.get('data', var_name))
 
-    nc_fo = os.path.join(root_dir, config.get('general', 'input_dir'), config.get('data', var_name))
-    
+    data_fo = os.path.join(root_dir, config.get('general', 'input_dir'), config.get('data', var_name)).rsplit(',')
+
[…]
