+++ example/example_settings.cfg
@@ -40,8 +40,8 @@ code2class=KoeppenGeiger/classification_codes.txt
 
 [data]
 # variable name here needs to be identical with variable name in nc-file
-total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc
-precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
+total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc,ln
+precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,ln
 temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc
 irr_water_demand=hydro/irrWaterDemand.nc
 
