### 2020-12-11 5f68bd6
log-transformed conflict count at t-1
