+++ copro/variables.py
@@ -130,9 +130,6 @@ def nc_with_continous_datetime_timestamp(extent_gdf, config, root_dir, var_name,
 
     if config.getboolean('general', 'verbose'): print('DEBUG: calculating mean {0} per aggregation unit from file {1} for year {2}'.format(var_name, nc_fo, sim_year))
 
-    print(nc_fo)
-    print(ln_flag)
-
     # open nc-file with xarray as dataset
     nc_ds = xr.open_dataset(nc_fo)
     # get xarray data-array for specified variable
