### 2020-12-11 d9f6220
masking negative and 0 values when log transforming
