### 2020-12-14 2915b36
different statistical methods for rasterstats
