### 2020-12-14 189dd93
different statistical methods for rasterstats
