### 2020-12-14 db5f614
optimemized workflow to log-transform values
