+++ example/example_settings.cfg
@@ -4,7 +4,7 @@ output_dir=./OUT
 # 1: all data; 2: leave-one-out model; 3: single variable model; 4: dubbelsteenmodel
 # Note that only 1 supports sensitivity_analysis
 model=1
-verbose=False
+verbose=True
 
 [settings]
 # start year
