### 2020-12-14 cde38e6
first implementatoin to find neighbors on run time
