### 2020-12-14 a9ff459
first implementatoin to find neighbors on run time
