### 2020-12-14 b380fcb
fully tested calculation of conflict at t-1 in poly neighbors
