+++ example/example_settings.cfg
@@ -45,8 +45,8 @@ code2class=KoeppenGeiger/classification_codes.txt
 # NOTE: only statistical functions supported by rasterstats are valid
 total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc,True,mean
 precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
-temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,False,mean
-irr_water_demand=hydro/irrWaterDemand.nc,False,sum
+temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
+irr_water_demand=hydro/irrWaterDemand.nc,True,mean
 
 [machine_learning]
 # choose from: MinMaxScaler, StandardScaler, RobustScaler, QuantileTransformer
