+++ example/example_settings_proj.cfg
@@ -43,10 +43,10 @@ code2class=KoeppenGeiger/classification_codes.txt
 # these three settings need to be separated by a comma
 # NOTE: variable name here needs to be identical with variable name in nc-file
 # NOTE: only statistical functions supported by rasterstats are valid
-total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc,True,mean
+total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc,False,mean
 precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
 temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,False,mean
-irr_water_demand=hydro/irrWaterDemand.nc,False,sum
+irr_water_demand=hydro/irrWaterDemand.nc,False,mean
 
 [machine_learning]
 # choose from: MinMaxScaler, StandardScaler, RobustScaler, QuantileTransformer
