+++ copro/conflict.py
@@ -91,8 +91,6 @@ def conflict_in_previous_year(config, conflict_gdf, extent_gdf, sim_year, neighb
     for i in range(len(extent_gdf)):
         i_poly = extent_gdf.watprovID.iloc[i]
         if i_poly in conflicts_per_poly.index.values:
-            # val = float(conflicts_per_poly.id.loc[conflicts_per_poly.index == i_poly].values[0])
-            # list_out.append(val)
             # find neighbors of this polygon
             nb = data.find_neighbors(i_poly, neighboring_matrix)
             # initialize count for total numbers of conflicts in neighbors
@@ -103,15 +101,15 @@ def conflict_in_previous_year(config, conflict_gdf, extent_gdf, sim_year, neighb
                 if k in conflicts_per_poly.index.values:
                     # determine number of conflicts per neigbors at t-1
                     val = conflicts_per_poly.id.loc[conflicts_per_poly.index == k].values[0]
-                    print('DEBUG: watprovID {} has neighbor {} with {} conflict(s) in previous timestep'.format(i_poly, k, val))
+                    if config.getboolean('general', 'verbose'): print('DEBUG: watprovID {} has neighbor {} with {} conflict(s) in previous timestep'.format(i_poly, k, val))
                     # add to sum
                     tot_nr_confl += val
             # append log-transformed sum to list
-            print('DEBUG: total number of conflicts at t-1 for watprovID {} is {}'.format(i_poly, tot_nr_confl))
+            if config.getboolean('general', 'verbose'): print('DEBUG: total number of conflicts at t-1 for watprovID {} is {}'.format(i_poly, tot_nr_confl))
             # if tot_nr_confl == 0.0: tot_nr_confl = 1.0
             tot_nr_confl = np.log(tot_nr_confl)
             if tot_nr_confl == -math.inf:
-                print('WARNING: no -inf allowed; setting value for watprovID {} to 0'.format(i_poly))
+                print('WARNING: no -inf allowed - setting value for watprovID {} to 0'.format(i_poly))
                 tot_nr_confl = 0.0
             list_out.append(tot_nr_confl)
         else:
