### 2020-12-16 7514fef
now both t-1 (bool) and t-1 for neihgbors (ln) sampled
