+++ copro/evaluation.py
@@ -323,7 +323,8 @@ def get_feature_importance(clf, config, out_dir):
     dict_out = dict()
     for key, x in zip(config.items('data'), range(len(arr))):
         dict_out[key[0]] = arr[x]
-    dict_out['conflict_t-1'] = arr[-1]
+    dict_out['conflict_t_min_1'] = arr[-2]
+    dict_out['conflict_t_min_1_nb'] = arr[-1]
 
     df = pd.DataFrame.from_dict(dict_out, orient='index', columns=['feature_importance'])
 
