+++ setup.py
@@ -12,7 +12,8 @@ requirements = ['Click>=7.0',
                 'rasterstats==0.14',
                 'geopandas==0.8.0',
                 'numpy==1.18.1',
-                'scikit-learn>=0.22.1',]
+                'scikit-learn>=0.22.1',
+                'seaborn==0.11']
 
 setup_requirements = ['pytest-runner', ]
 
