### 2020-12-17 0f444e7
now boolean instead of ln-value for conflict in nb polys
