### 2020-12-17 5eaf559
now boolean instead of ln-value for conflict in nb polys
