### 2020-12-17 1b3d2d2
dumping and loading clfs per run via pickle
