### 2020-12-17 d0ab000
first step to store bool conflict map at last ref run year
