### 2020-12-17 7372003
making output folder for (temp) files
