+++ copro/data.py
@@ -58,7 +58,7 @@ def initiate_X_data(config):
 
     return X
 
-def fill_XY(XY, config, root_dir, conflict_gdf, polygon_gdf):
+def fill_XY(XY, config, root_dir, conflict_gdf, polygon_gdf, out_dir):
     """Fills the XY-dictionary with data for each variable and conflict for each polygon for each simulation year. 
     The number of rows should therefore equal to number simulation years times number of polygons.
     At end of last simulation year, the dictionary is converted to a numpy-array.
@@ -100,7 +100,7 @@ def fill_XY(XY, config, root_dir, conflict_gdf, polygon_gdf):
                 if key == 'conflict':
                 
                     data_series = value
-                    data_list = conflict.conflict_in_year_bool(config, conflict_gdf, polygon_gdf, sim_year)
+                    data_list = conflict.conflict_in_year_bool(config, conflict_gdf, polygon_gdf, sim_year, out_dir)
                     data_series = data_series.append(pd.Series(data_list), ignore_index=True)
                     XY[key] = data_series
 
