### 2020-12-18 #2855 @sbenthall
I think this is terribly exciting and I am eager to see it published in the best form possible. To that end:

* Please document the requirement that Conda needs to be installed to run the installation instructions.
* Is there any other comparable software that should be considered as the 'state of the field' in this line of research?
* I understand that the library has been designed for flexibility. However, I am having trouble determining from the paper and indeed from the documentation what the core functionality of the library is. I wonder if it would be possible to foreground a bit more technical detail about what scaffolding the library is doing for modeling work, and what kinds of models are included out-of-the-box.

I have not yet completed reviewing the code--these are first impressions of the document and paper only. But I think these improvements to the writeup and documentation are important for the publication.
