### 2020-12-18 #2855 @sbenthall
When I get to the `sh run.sh` part of the instructions in the README, I get the following error:

https://gist.github.com/sbenthall/83de0ce8675283c01d3f3fa973dc6f38
