### 2020-12-20 #2855 @JannisHoch
Hi @sbenthall,
Absolutely no problem!

From your comments I assume you were not looking at the most recent version of the model.
For instance, in the joss-review branch (https://github.com/JannisHoch/copro/tree/joss_review), there is no file run.sh anymore. It is renamed to 'run_notebooks.sh' for running the jupyter notebooks, and 'run_script_*.sh' for executing the command line scripts.

If your problems still occur after installing the latest copro version, please let me know. thanks :)
