### 2020-12-20 d68c515
still saving clf fitted with all samples until new approach works
