### 2020-12-20 7b1247d
not using conflict at t-1 for projections until new approach works
