+++ copro/machine_learning.py
@@ -162,13 +162,13 @@ def pickle_clf(scaler, clf, config, root_dir):
 
     X_fit, Y_fit = data.split_XY_data(XY_fit, config)
     X_ID_fit, X_geom_fit, X_data_fit = conflict.split_conflict_geom_data(X_fit)
-    X_ft_fit = scaler.fit_transform(X_data_fit[:, :-3])
+    X_ft_fit = scaler.fit_transform(X_data_fit)
 
     clf.fit(X_ft_fit, Y_fit)
 
-    print('INFO: dumping classifier to {}'.format(os.path.join(root_dir, config.get('general', 'output_dir'), 'clf.pkl')))
-    with open(os.path.join(root_dir, config.get('general', 'output_dir'), 'clf.pkl'), 'wb') as f:
-        pickle.dump(clf, f)
+    # print('INFO: dumping classifier to {}'.format(os.path.join(root_dir, config.get('general', 'output_dir'), 'clf.pkl')))
+    # with open(os.path.join(root_dir, config.get('general', 'output_dir'), 'clf.pkl'), 'wb') as f:
+    #     pickle.dump(clf, f)
 
     return clf
 
