### 2020-12-20 5ddee78
Update README.rst
added note and link to anaconda/miniconda installation
