### 2020-12-20 f9326d3
included link to anaconda/miniconda
