### 2020-12-20 #2855 @JannisHoch
in f9326d36bcbf20e93a54e6ea5733f1ae515f4477 I have added notes that the model was developed and tested with anaconda/miniconda.
