### 2020-12-20 #2855 @JannisHoch
regarding your question of other state-of-the-art software: to my knowledge, there is no freely available software model with such an extensive documentation out there. of course there is model code as supplement to a publication, and publications covering similar approaches (see the references in the joss manuscript), but not in a 'structured' way like done for CoPro.
