+++ joss/paper.md
@@ -34,13 +34,13 @@ In light of predicted shifts in climate patterns and the potentially resulting b
 
 # Statement of need 
 
-There is increasing consensus that climate change can exacerbate the risk of (armed) conflict [@koubi2019climate; @mach2019climate]. Nevertheless, making (operational) forecasts on the short-term is still challenging due to several reasons [@cederman2017predicting]. Building upon recent, similar approaches to use data-driven models [@colaresi2017robot] and statistical approaches [@witmer2017subnational; @hegre2016forecasting] for conflict risk projections, CoPro is a novel, fully open, and extensible Python-model facilitating the set-up, execution, and evaluation of machine-learning models predicting conflict risk. CoPro provides a structured workflow including pre- and post-processing tools, making it accessible to all levels of experience. Such a user-friendly tool is needed not only to integrate the different disciplines, but also to extend the modeling approach with new insights and data - after all, the established links between climate and societal factors with conflict are still weak [@koubi2019climate; @mach2019climate]. In addition to scholarly explorations of the inter-dependencies and importance of various conflict drivers, model output such as maps of spatially-disaggregated projected conflict risk can be an invaluable input to inform the decision-making process in affected regions.
+There is increasing consensus that climate change can exacerbate the risk of (armed) conflict [@koubi2019climate; @mach2019climate]. Nevertheless, making (operational) forecasts on the short-term is still challenging due to several reasons [@cederman2017predicting]. Building upon recent, similar approaches to use data-driven models [@colaresi2017robot] and statistical approaches [@witmer2017subnational; @hegre2016forecasting] for conflict risk projections, CoPro is a novel, fully open, and extensible Python-model facilitating the set-up, execution, and evaluation of machine-learning models predicting conflict risk. CoPro provides a structured workflow including pre- and post-processing tools, making it accessible to all levels of experience. Such a user-friendly tool is needed not only to integrate the different disciplines, but also to extend the modeling approach with new insights and data - after all, the established links between climate and societal factors with conflict are still weak [@koubi2019climate; @mach2019climate]. In addition to scholarly explorations of the inter-dependencies and the importance of various conflict drivers, model output such as maps of spatially-disaggregated projected conflict risk can be an invaluable input to inform the decision-making process in affected regions.
 
-Since conflicts are of all times and not limited to specific regions or countries, CoPro is designed with user-flexibility in mind. Therefore, the number and variables provided to the model is not specified, allowing for bespoke model designs. Depending on the modeling exercise and data used, several machine-learning models and pre-processing algorithms are available in CoPro. Catering for different model designs is of added value because of the non-linear and sometimes irrational - 'law-breaking' [@cederman2017predicting] - nature of conflicts. On top of that, the analyses can be run at any spatial scale, allowing for better identification of sub-national drivers of conflict risk. After all, conflict onset and conflicts are often limited to specific areas where driving factors coincide. 
+Since conflicts are of all times and not limited to specific regions or countries, CoPro is designed with user-flexibility in mind. Therefore, the number and variables provided to the model is not specified, allowing for bespoke model designs. Depending on the modeling exercise and data used, several machine-learning models and pre-processing algorithms are available in CoPro. In its current form, the supervised learning techniques support vector classifier, k-neighbors classifier, and random-forest classifier are implemented. Catering for different model designs is of added value because of the non-linear and sometimes irrational - 'law-breaking' [@cederman2017predicting] - nature of conflicts. On top of that, the analyses can be run at any spatial scale, allowing for better identification of sub-national drivers of conflict risk. After all, conflict onset and conflicts are often limited to specific areas where driving factors coincide. 
 
[…]
