### 2020-12-20 #2855 @JannisHoch
I also have added a bit more technical detail to the JOSS manuscript. based on the review of @soodoku, the model goal and purpose were made clearer in the README and in the documentation.
