### 2020-12-20 6dda5f7
towards v0.0.8 (i.e. after JOSS publication version)
