+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.0.7b
+current_version = 0.0.8b
 commit = True
 tag = True
 
