### 2020-12-20 3ee06a5
towards v0.0.8 (i.e. after JOSS publication version)
