+++ copro/__init__.py
@@ -13,4 +13,4 @@ from . import plots
 
 __author__ = """Jannis M. Hoch, Sophie de Bruin, Niko Wanders"""
 __email__ = 'j.m.hoch@uu.nl'
-__version__ = '0.0.7b'
+__version__ = '0.0.8b'
