### 2020-12-20 fb85bb8
storing boolean conflict map to file
