+++ example/_scripts/run_script_projections.sh
@@ -1,0 +1,1 @@
+copro_runner ../example_settings.cfg -proj ../example_settings_proj.cfg
