+++ example/run_script_reference.sh
@@ -1,1 +1,0 @@
-copro_runner example_settings.cfg
