+++ example/run_script_projections.sh
@@ -1,1 +1,0 @@
-copro_runner example_settings.cfg -proj example_settings_proj.cfg
