### 2021-01-08 3406d4d
storing binary conflicht map in temp_folder for notebooks
