+++ example/_scripts/run_notebooks.sh
@@ -1,6 +1,6 @@
 jupyter nbconvert --to html --ExecutePreprocessor.allow_errors=True --ExecutePreprocessor.timeout=-1 --ExecutePreprocessor.kernel_name="python3" --execute ../*.ipynb
 
 echo removing temporary files
-rm ../conflicts.*
-rm ../polygons.*
-rm ../global_df.npy
+rm ../temp_files/conflicts.*
+rm ../temp_files/polygons.*
+rm ../temp_files/global_df.npy
