### 2021-01-11 ea00ef9
updating some fucntion doc
