### 2021-01-11 c384877
testing new function to parse cfg-files per projection
