+++ example/example_settings.cfg
@@ -6,21 +6,21 @@ output_dir=./OUT
 model=1
 verbose=True
 
-[settings]
+[REF_settings]
 # start year
 y_start=2000
 # end year
 y_end=2015
-# number of repetitions
-n_runs=10
+
+[PROJ_files]
+# cfg-files
+test1=./example_settings_proj.cfg
+test_blub=./example_settings_proj.cfg
 
 [pre_calc]
 # if nothing is specified, the XY array will be stored in output_dir
 # if XY already pre-calculated, then provide path to npy-file
 XY=
-# if nothing is specified, the classifier will be stored in output_dir
-# if classifier is already stored, then provide path to pkl-file
-clf=
 
 [extent]
 shp=waterProvinces/waterProvinces_Africa.shp
@@ -54,3 +54,5 @@ scaler=QuantileTransformer
 # choose from: NuSVC, KNeighborsClassifier, RFClassifier
 model=RFClassifier
 train_fraction=0.7
+# number of repetitions
+n_runs=10
