### 2021-01-11 181814d
testing new function to parse cfg-files per projection
