+++ example/example_settings_proj.cfg
@@ -1,42 +1,13 @@
 [general]
-input_dir=./example_data
-output_dir=./OUT_PROJ
-# 1: all data; 2: leave-one-out model; 3: single variable model; 4: dubbelsteenmodel
-# Note that only 1 supports sensitivity_analysis
-model=1
-verbose=True
-
-[settings]
 # start year
 y_start=2010
 # end year
 y_end=2015
-# number of repetitions
-n_runs=50
 
 [pre_calc]
 # if nothing is specified, the XY array will be stored in output_dir
 # if XY already pre-calculated, then provide (absolute) path to npy-file
 XY=
-# if nothing is specified, the classifier will be stored in output_dir
-# if classifier is already stored, then provide (absolute) path to pkl-file
-clf=./OUT/clf.pkl
-
-[extent]
-shp=waterProvinces/waterProvinces_Africa.shp
-
-[conflict]
-# either specify path to file or state 'download' to download latest PRIO/UCDP dataset
-conflict_file=UCDP/ged201.csv
-min_nr_casualties=1
-# 1=state-based armed conflict; 2=non-state conflict; 3=one-sided violence
-type_of_violence=1,2,3
-
-[climate]
-shp=KoeppenGeiger/2000/Koeppen_Geiger_1976-2000.shp
-# define either one or more classes (use abbreviations!) or specify None for not filtering
-zones=BWh,BSh
-code2class=KoeppenGeiger/classification_codes.txt
 
 [data]
 # specify the path to the nc-file, whether the variable shall be log-transformed (True, False), and which statistical function should be applied
@@ -47,10 +18,3 @@ total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearme
 precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
 temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,False,mean
 irr_water_demand=hydro/irrWaterDemand.nc,False,mean
-
-[machine_learning]
-# choose from: MinMaxScaler, StandardScaler, RobustScaler, QuantileTransformer
-scaler=QuantileTransformer
-# choose from: NuSVC, KNeighborsClassifier, RFClassifier
-model=RFClassifier
-train_fraction=0.7
