### 2021-01-11 a86cbfc
new output folder structure implemented
