### 2021-01-11 d2ea489
new output folder structure implemented
