### 2021-01-11 b943a5c
minor changes in function design
