+++ copro/utils.py
@@ -109,9 +109,6 @@ def parse_projection_settings(config):
         each_val = os.path.abspath(each_val)
 
         # parse each config-file specified
-        # each_config = RawConfigParser(allow_no_value=True, inline_comment_prefixes='#')
-        # each_config.optionxform = lambda option: option
-        # each_config.read(each_val)
         each_config = parse_settings(each_val)
 
         # update the output dictionary with key and config-object
@@ -128,7 +125,7 @@ def make_output_dir(config, root_dir, proj_settings_dir):
         proj_settings_dir (dict): dictionary containing config-objects per projection.
 
     Returns:
-        str: path to output folder
+        list: list with output directories, first entry refers to main dir, second to reference situation, all following to each projection run.
     """    
 
     out_dir = os.path.join(root_dir, config.get('general','output_dir'))
@@ -136,31 +133,30 @@ def make_output_dir(config, root_dir, proj_settings_dir):
 
     out_dir_list = []
 
+    out_dir_list.append(out_dir)
     out_dir_list.append(os.path.join(out_dir, '_REF'))
 
     out_dir_proj = os.path.join(out_dir, '_PROJ')
     for key in proj_settings_dir:
         out_dir_list.append(os.path.join(out_dir_proj, str(key)))
-    
-    for d in out_dir_list:
-        print('DEBUG: ...and sub-directory {}'.format(d))
 
-    # if not os.path.isdir(out_dir):
-    #     os.makedirs(out_dir)
-    #     os.makedirs(os.path.join(out_dir, '_REF'))
-    #     os.makedirs(os.path.join(out_dir, '_PROJ'))
-    # else:
-    #     for root, dirs, files in os.walk(out_dir):
-    #         if config.getboolean('general', 'verbose'): print('DEBUG: remove files in {}'.format(os.path.abspath(root)))
-    #         for fo in files:
-    #             # print(fo)
-    #             if (fo =='XY.npy') or (fo == 'X.npy'):
-    #                 if config.getboolean('general', 'verbose'): print('DEBUG: sparing {}'.format(fo))
-    #                 pass
-    #             else:
-    #                 os.remove(os.path.join(root, fo))
+    for d in out_dir_list:
+        
+        if not os.path.isdir(d):
+            print('INFO: creating output-folder {}'.format(d))
+            os.makedirs(d)
+
+        else:
+            for root, dirs, files in os.walk(d):
+                if config.getboolean('general', 'verbose'): print('DEBUG: remove files in {}'.format(os.path.abspath(root)))
+                for fo in files:
+                    if (fo =='XY.npy') or (fo == 'X.npy'):
+                        if config.getboolean('general', 'verbose'): print('DEBUG: sparing {}'.format(fo))
+                        pass
+                    else:
+                        os.remove(os.path.join(root, fo))
                             
-    return out_dir
+    return out_dir_list
     
 def download_PRIO(config, root_dir):
     """If specfied in cfg-file, the PRIO/UCDP data is directly downloaded and used as model input.
@@ -228,18 +224,19 @@ def initiate_setup(settings_file):
 
     print('INFO: verbose mode on: {}'.format(config.getboolean('general', 'verbose')))
 
-    out_dir = make_output_dir(config, root_dir, proj_settings_dir)
+    out_dir_list = make_output_dir(config, root_dir, proj_settings_dir)
 
-    copyfile(settings_file, os.path.join(out_dir, 'copy_of_run_setting.cfg'))
+    print('DEBUG: copying cfg-file {} to folder {}'.format(settings_file, out_dir_list[0]))
+    copyfile(settings_file, os.path.join(out_dir_list[0], 'copy_of_run_setting.cfg'))
 
[…]
