### 2021-01-11 fe434c1
make changes work for reference run notebooks
