### 2021-01-11 5951140
make changes work for reference run notebooks
