+++ copro/conflict.py
@@ -37,7 +37,7 @@ def conflict_in_year_bool(config, conflict_gdf, extent_gdf, sim_year, out_dir):
     if not os.path.isdir(out_dir):
         os.makedirs(out_dir)
 
-    if sim_year == config.getint('settings', 'y_end'):
+    if sim_year == config.getint('REF_settings', 'y_end'):
         if config.getboolean('general', 'verbose'): print('DEBUG: storing boolean conflict map of year {} to file {}'.format(sim_year, os.path.join(out_dir, 'conflicts_in_{}.shp'.format(sim_year))))
         # get a 1 for each polygon where there was conflict
         bool_per_poly = fatalities_per_poly / fatalities_per_poly
