### 2021-01-11 a8abeb5
minor update in print statements
