### 2021-01-11 fd388c3
working with dict of config-objects
