### 2021-01-11 e4de057
merging config-ojbects and out dirs in one dict
