+++ example/example_settings.cfg
@@ -6,7 +6,7 @@ output_dir=./OUT
 model=1
 verbose=True
 
-[REF_settings]
+[settings]
 # start year
 y_start=2000
 # end year
@@ -15,7 +15,6 @@ y_end=2015
 [PROJ_files]
 # cfg-files
 test1=./example_settings_proj.cfg
-test_blub=./example_settings_proj.cfg
 
 [pre_calc]
 # if nothing is specified, the XY array will be stored in output_dir
