### 2021-01-11 a3d9615
merging config-ojbects and out dirs in one dict
