+++ example/example_settings_proj.cfg
@@ -1,4 +1,7 @@
 [general]
+verbose=True
+
+[settings]
 # start year
 y_start=2010
 # end year
