### 2021-01-13 a001840
minor bugfix in cfg-section name
