### 2021-01-13 b2ef696
making independent projectoins per clf
