+++ example/example_settings_proj.cfg
@@ -1,9 +1,10 @@
 [general]
+input_dir=./example_data
 verbose=True
 
 [settings]
 # start year
-y_start=2010
+y_start=2013
 # end year
 y_end=2015
 
