### 2021-01-13 3866dee
enabled making various projs + apply clfs per proj
