+++ example/example_settings.cfg
@@ -15,6 +15,7 @@ y_end=2015
 [PROJ_files]
 # cfg-files
 test1=./example_settings_proj.cfg
+testZWEI=./example_settings_proj.cfg
 
 [pre_calc]
 # if nothing is specified, the XY array will be stored in output_dir
