### 2021-01-13 37c61c3
enabled making various projs + apply clfs per proj
