+++ example/example_settings.cfg
@@ -8,9 +8,9 @@ verbose=True
 
 [settings]
 # start year
-y_start=2000
+y_start=1995
 # end year
-y_end=2015
+y_end=2014
 
 [PROJ_files]
 # cfg-files
@@ -43,10 +43,11 @@ code2class=KoeppenGeiger/classification_codes.txt
 # these three settings need to be separated by a comma
 # NOTE: variable name here needs to be identical with variable name in nc-file
 # NOTE: only statistical functions supported by rasterstats are valid
-total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc,True,mean
-precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
-temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
-irr_water_demand=hydro/irrWaterDemand.nc,True,mean
+# total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc,True,mean
+# precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
+# temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
+# irr_water_demand=hydro/irrWaterDemand.nc,True,mean
+gdp=gdp/gdp_historic_5min_Africa.nc,True,mean
 
 [machine_learning]
 # choose from: MinMaxScaler, StandardScaler, RobustScaler, QuantileTransformer
