### 2021-01-14 e4f5394
defining gap period between REF and PROJ
