+++ example/example_settings_proj.cfg
@@ -4,9 +4,9 @@ verbose=True
 
 [settings]
 # start year
-y_start=2013
+y_start=2020
 # end year
-y_end=2015
+y_end=2025
 
 [pre_calc]
 # if nothing is specified, the XY array will be stored in output_dir
@@ -18,7 +18,8 @@ XY=
 # these three settings need to be separated by a comma
 # NOTE: variable name here needs to be identical with variable name in nc-file
 # NOTE: only statistical functions supported by rasterstats are valid
-total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc,False,mean
-precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
-temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,False,mean
-irr_water_demand=hydro/irrWaterDemand.nc,False,mean
+# total_evaporation=hydro/totalEvaporation_monthTot_output_2000_2015_Africa_yearmean.nc,False,mean
+# precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
+# temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,False,mean
+# irr_water_demand=hydro/irrWaterDemand.nc,False,mean
+gdp=gdp/gdp_SSP1_5min_Africa.nc,True,mean
