### 2021-01-14 de9af68
dynamic calc of gap period per PROJ + out_dir
