+++ copro/models.py
@@ -151,19 +151,36 @@ def dubbelsteen(X, Y, config, scaler, clf, out_dir):
 
     return X_df, y_df, eval_dict
 
-def fill_gap_period(config_REF, config_PROJ, out_dir_PROJ):
+def determine_projection_period(config_REF, config_PROJ, out_dir_PROJ):
 
-    print('INFO: determinining conflict occurence for gap period between reference and projection run')
+    print('INFO: determinining annual conflict occurence from end of reference run until end of projection run')
 
     if not os.path.isdir(os.path.join(out_dir_PROJ, 'files')):
         print('DEBUG: creating output folder for annual conflict maps {}'.format(os.path.join(out_dir_PROJ, 'files')))
         os.makedirs(os.path.join(out_dir_PROJ, 'files'))
 
-    gap_period = np.arange(config_REF.getint('settings', 'y_end')+1, config_PROJ.getint('settings', 'y_start'), 1)
-    gap_period = gap_period.tolist()
-    print('DEBUG: the gap period is {}'.format(gap_period))
+    projection_period = np.arange(config_REF.getint('settings', 'y_end')+1, config_PROJ.getint('settings', 'y_end')+1, 1)
+    projection_period = projection_period.tolist()
+    print('DEBUG: the projection period is {}'.format(projection_period))
 
-    return gap_period
+    return projection_period
+
+def fill_projection_period(config_REF, out_dir_REF, config_PROJ, out_dir_PROJ):
+
+    projection_period = determine_projection_period(config_REF, config_PROJ, out_dir_PROJ)
+
+    for i in range(len(projection_period)):
+        print('INFO: entering year {} in projection period'.format(projection_period[i]))
+        if i == 0:
+            if os.path.isfile(os.path.join(out_dir_REF, 'files', 'conflicts_in_{}.csv'.format(config_REF.getint('settings', 'y_end')))):
+                print('DEBUG: reading from last conflict observation {}'.format(os.path.join(out_dir_REF, 'files', 'conflicts_in_{}.csv'.format(config_REF.getint('settings', 'y_end')))))
+                conflict_gdf = pd.read_csv(os.path.join(out_dir_REF, 'files', 'conflicts_in_{}.csv'.format(config_REF.getint('settings', 'y_end'))), index_col=0)
+            else:
+                raise ValueError('ERROR: the file with binary conflict occurence at last time step of reference run could not be found!')
+        else:
+            conflict_gdf = None
+
+    return projection_period
 
 def predictive(X, scaler, main_dict, root_dir):
     """Predictive model to use the already fitted classifier to make projections.
