### 2021-01-14 a35c0e2
unifying different simulation periods for proj
