### 2021-01-14 a24bd2d
towards loading previous conflict data
