+++ example/example_settings_proj.cfg
@@ -22,4 +22,4 @@ XY=
 # precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
 # temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,False,mean
 # irr_water_demand=hydro/irrWaterDemand.nc,False,mean
-gdp=gdp/gdp_SSP1_5min_Africa.nc,True,mean
+gdp=gdp/gdp_Africa.nc,True,mean
