+++ example/example_settings.cfg
@@ -10,7 +10,7 @@ verbose=True
 # start year
 y_start=1995
 # end year
-y_end=2014
+y_end=2012
 
 [PROJ_files]
 # cfg-files
@@ -47,7 +47,7 @@ code2class=KoeppenGeiger/classification_codes.txt
 # precipitation=hydro/precipitation_monthTot_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
 # temperature=hydro/temperature_monthAvg_output_2000-01-31_to_2015-12-31_Africa_yearmean.nc,True,mean
 # irr_water_demand=hydro/irrWaterDemand.nc,True,mean
-gdp=gdp/gdp_historic_5min_Africa.nc,True,mean
+gdp=gdp/gdp_Africa.nc,True,mean
 
 [machine_learning]
 # choose from: MinMaxScaler, StandardScaler, RobustScaler, QuantileTransformer
