### 2021-01-14 f37ae3f
towards loading previous conflict data
