+++ copro/conflict.py
@@ -5,7 +5,7 @@ import numpy as np
 import os, sys
 import math
 
-def conflict_in_year_bool(config, conflict_gdf, extent_gdf, sim_year, out_dir, proj=False): 
+def conflict_in_year_bool(config, conflict_gdf, extent_gdf, sim_year, out_dir): 
     """Creates a list for each timestep with boolean information whether a conflict took place in a polygon or not.
 
     Args:
@@ -20,6 +20,8 @@ def conflict_in_year_bool(config, conflict_gdf, extent_gdf, sim_year, out_dir, p
         list: list containing 0/1 per polygon depending on conflict occurence.
     """    
 
+    if config.getboolean('general', 'verbose'): print('DEBUG: checking for conflict event in polygon at t')
+
     # select the entries which occured in this year
     temp_sel_year = conflict_gdf.loc[conflict_gdf.year == sim_year]  
 
@@ -27,7 +29,7 @@ def conflict_in_year_bool(config, conflict_gdf, extent_gdf, sim_year, out_dir, p
     
     # merge the dataframes with polygons and conflict information, creating a sub-set of polygons/regions
     data_merged = gpd.sjoin(temp_sel_year, extent_gdf)
-    
+
     # determine the aggregated amount of fatalities in one region (e.g. water province)
     try:
         fatalities_per_poly = data_merged['best'].groupby(data_merged['watprovID']).sum().to_frame().rename(columns={"best": 'total_fatalities'})
@@ -51,7 +53,10 @@ def conflict_in_year_bool(config, conflict_gdf, extent_gdf, sim_year, out_dir, p
         # merge the boolean info with geometry
         # for all polygons without conflict, set a 0
         if config.getboolean('general', 'verbose'): print('DEBUG: storing boolean conflict map of year {} to file {}'.format(sim_year, os.path.join(out_dir, 'conflicts_in_{}.csv'.format(sim_year))))
-        data_stored = pd.merge(bool_per_poly, global_df, on='ID', how='right').fillna(0)
+        # data_stored = pd.merge(bool_per_poly, global_df, on='ID', how='right').fillna(0)
+        data_stored = pd.merge(bool_per_poly, global_df, on='ID', how='right').dropna()
+        data_stored.index = data_stored.index.rename('watprovID')
+        data_stored = data_stored.drop('geometry', axis=1)
         data_stored.to_csv(os.path.join(out_dir, 'conflicts_in_{}.csv'.format(sim_year)))
  
     # loop through all regions and check if exists in sub-set
@@ -67,12 +72,11 @@ def conflict_in_year_bool(config, conflict_gdf, extent_gdf, sim_year, out_dir, p
         else:
             list_out.append(0)
             
-    if not len(extent_gdf) == len(list_out):
-        raise AssertionError('the dataframe with polygons has a lenght {0} while the lenght of the resulting list is {1}'.format(len(extent_gdf), len(list_out)))
+    assert (len(extent_gdf) == len(list_out)), AssertionError('the dataframe with polygons has a lenght {0} while the lenght of the resulting list is {1}'.format(len(extent_gdf), len(list_out)))
 
     return list_out
 
-def conflict_in_previous_year(config, conflict_gdf, extent_gdf, sim_year, check_neighbors=False, neighboring_matrix=None, proj=False):
+def conflict_in_previous_year(config, conflict_gdf, extent_gdf, sim_year, check_neighbors=False, neighboring_matrix=None):
     """Creates a list for each timestep with boolean information whether a conflict took place in a polygon at the previous timestep or not.
     If the current time step is the first (t=0), then conflict data of this year is used instead due to the lack of earlier data.
 
[…]
