### 2021-01-14 986727b
outputting evaluatoin metrics
