### 2021-01-14 40d55ac
layout how to make proj per clf per timestep
