### 2021-01-14 7eec6cf
layout how to make proj per clf per timestep
