+++ copro/models.py
@@ -181,9 +181,6 @@ def predictive(X, scaler, main_dict, root_dir):
         datatrame: containing model output on polygon-basis.
     """    
 
-    config_REF = main_dict['_REF'][0]
-    out_dir_REF = main_dict['_REF'][1]
-
     print('INFO: scaling the data from projection period')
     X = pd.DataFrame(X)
     if config_REF.getboolean('general', 'verbose'): print('DEBUG: number of data points including missing values: {}'.format(len(X)))
@@ -193,19 +190,10 @@ def predictive(X, scaler, main_dict, root_dir):
     ##- scaling only the variable values
     X_ft = scaler.fit_transform(X_data)
 
-    clfs = machine_learning.load_clfs(config_REF, out_dir_REF)
-
-    y_df = pd.DataFrame(columns=['ID', 'geometry', 'y_pred'])
-
     print('INFO: making the projections')    
-    for clf in clfs:    
 
-        with open(os.path.join(out_dir_REF, 'clfs', clf), 'rb') as f:
-            print('DEBUG: loading classifier {} from {}'.format(clf, os.path.join(out_dir_REF, 'clfs')))
-            clf = pickle.load(f)
-
-        y_pred = clf.predict(X_ft)
-        arr = np.column_stack((X_ID, X_geom, y_pred))
-        y_df = y_df.append(pd.DataFrame(arr, columns=['ID', 'geometry', 'y_pred']), ignore_index=True)
-
+    y_pred = clf.predict(X_ft)
+    arr = np.column_stack((X_ID, X_geom, y_pred))
+    y_df = pd.DataFrame(arr, columns=['ID', 'geometry', 'y_pred'])
+    
     return y_df
