### 2021-01-15 c048a09
making proj per time step, still with not-updated conflicat data
