### 2021-01-15 be4df34
making proj per time step, still with not-updated conflicat data
