+++ example/example_settings_proj.cfg
@@ -3,10 +3,8 @@ input_dir=./example_data
 verbose=True
 
 [settings]
-# start year
-y_start=2020
-# end year
-y_end=2025
+# year for which projection is to be made
+y_proj=2025
 
 [pre_calc]
 # if nothing is specified, the XY array will be stored in output_dir
