### 2021-01-18 37bbb9d
checking code in notebooks
