### 2021-01-18 3373aeb
cl script working for ref run
