+++ copro/scripts/copro_runner.py
@@ -11,10 +11,10 @@ warnings.filterwarnings("ignore")
 
 @click.command()
 @click.argument('cfg', type=click.Path())
-@click.option('--projection-settings', '-proj', help='path to cfg-file with settings for a projection run', multiple=True, type=click.Path())
+@click.option('--make_plots', '-plt', help='add additional output plots')
 @click.option('--verbose', '-v', help='command line switch to turn on verbose mode', is_flag=True)
 
-def cli(cfg, projection_settings=[], verbose=False):   
+def cli(cfg, make_plots=True, verbose=False):   
     """Main command line script to execute the model. 
     All settings are read from cfg-file.
     One cfg-file is required argument to train, test, and evaluate the model.
@@ -25,27 +25,32 @@ def cli(cfg, projection_settings=[], verbose=False):
     """ 
   
     #- parsing settings-file and getting path to output folder
-    config, out_dir, root_dir = copro.utils.initiate_setup(cfg)
+    main_dict, root_dir = copro.utils.initiate_setup(cfg)
+
+    config_REF = main_dict['_REF'][0]
+    out_dir_REF = main_dict['_REF'][1]
 
     if verbose:
-        config.set('general', 'verbose', str(verbose))
+        config_REF.set('general', 'verbose', str(verbose))
 
     click.echo(click.style('\nINFO: reference run started\n', fg='cyan'))
 
     #- selecting conflicts and getting area-of-interest and aggregation level
-    conflict_gdf, extent_gdf, extent_active_polys_gdf, global_df = copro.selection.select(config, out_dir, root_dir)
-    #- plot selected polygons and conflicts
-    fig, ax = plt.subplots(1, 1)
-    copro.plots.selected_polygons(extent_active_polys_gdf, figsize=(20, 10), ax=ax)
-    copro.plots.selected_conflicts(conflict_gdf, ax=ax)
-    plt.savefig(os.path.join(out_dir, 'selected_polygons_and_conflicts.png'), dpi=300, bbox_inches='tight')
+    conflict_gdf, extent_gdf, extent_active_polys_gdf, global_df = copro.selection.select(config_REF, out_dir_REF, root_dir)
+
+    if make_plots:
+        #- plot selected polygons and conflicts
+        fig, ax = plt.subplots(1, 1)
+        copro.plots.selected_polygons(extent_active_polys_gdf, figsize=(20, 10), ax=ax)
+        copro.plots.selected_conflicts(conflict_gdf, ax=ax)
+        plt.savefig(os.path.join(out_dir_REF, 'selected_polygons_and_conflicts.png'), dpi=300, bbox_inches='tight')
 
     #- create X and Y arrays by reading conflict and variable files;
     #- or by loading a pre-computed array (npy-file)
-    X, Y = copro.pipeline.create_XY(config, out_dir, root_dir, extent_active_polys_gdf, conflict_gdf)
+    X, Y = copro.pipeline.create_XY(config_REF, out_dir_REF, root_dir, extent_active_polys_gdf, conflict_gdf)
 
     #- defining scaling and model algorithms
-    scaler, clf = copro.pipeline.prepare_ML(config)
+    scaler, clf = copro.pipeline.prepare_ML(config_REF)
 
     #- initializing output variables
     #TODO: put all this into one function
@@ -59,12 +64,12 @@ def cli(cfg, projection_settings=[], verbose=False):
 
     click.echo('INFO: training and testing machine learning model')
     #- go through all n model executions
-    for n in range(config.getint('settings', 'n_runs')):
[…]
