### 2021-01-18 #2855 @sbenthall
I am being held up by some local Python environment issues. (The `rasterio` dependency is not installing properly.) I'm pretty sure this is just a problem on my end! I'll try to repair my environment setup and get back to you.
