### 2021-01-19 #2855 @sbenthall
This is the stack trace. https://gist.github.com/sbenthall/5edad4e128fea79a99c2486ab6333bfe
Googling for it makes it look like a local python issue. I'll keep kicking at it.
