### 2021-01-19 #2855 @sbenthall
Hmm.
So what's strange here is that `conda install rasterio==1.1.0` works fine, but then I get an error when it installs again via `pip install -e .`.

I wonder if the conda version is conflicting with the pip version somehow (the pip version is invoked in the `setup.py` file.)
