### 2021-01-19 #2855 @sbenthall
Commenting out the `rasterio` dependency in `setup.py` and rerunning the `pip` installation gets me this error:

https://gist.github.com/sbenthall/6a7f135970232dad08eb65ecebfd97cf
