### 2021-01-19 #2855 @sbenthall
I do notice that `rasterio` installation instructions look like they require a number of other GDAL related packages, which in the `rasterio` installation instructions are listed as separate steps.

https://rasterio.readthedocs.io/en/latest/

I wonder if you could confirm that your instructions for CoPro, as written, work on a fresh system that does not have these GDAL packages installed. It is quite likely that these are being handled correctly by Anaconda and there's some issue going on locally, but if you could test on another machine that would be additional confirmation that it's a local issue and not a problem with the installation instructions.
