+++ example/example_settings_proj.cfg
@@ -4,7 +4,7 @@ verbose=True
 
 [settings]
 # year for which projection is to be made
-y_proj=2025
+y_proj=2015
 
 [pre_calc]
 # if nothing is specified, the XY array will be stored in output_dir
