### 2021-01-20 #2855 @JannisHoch
Hmmm, this all sounds pretty odd to me.

I have never seen any of the error messages before when installing copro - neither the one when running `pip install -e .` (as it seems that rasterio is correctly installed from cache and the error occurs bit later somehow), nor the one with altered `setup.py` file (which is a permission error after all if I get it right?).

Also, I was able to install copro (branch joss_review) on a pretty 'fresh' Linux environment.

So yes, it may be that it's a local issue. I have two ideas:

- use a different linux/windows enviroment?
- in the `environment.yml` file, put the rasterio dependency in the pip section instead the conda-forge section and see if this solves the suspected conflict between conda/pip?
