### 2021-01-20 #2855 @JannisHoch
Or, as another possible fix, specify a more recent rasterio version? current one is 1.1.8. maybe there are some flaws in version 1.1.0 for some specific circumstances.
