### 2021-01-20 2a5e1e8
updated docs of command lien script
