### 2021-01-20 d6d9526
using fitted scaler for projections
