### 2021-01-20 004ecf4
using fitted scaler for projections
