### 2021-01-20 d961095
update docs for data.py
