+++ copro/data.py
@@ -79,7 +79,7 @@ def fill_XY(XY, config, root_dir, conflict_data, polygon_gdf, out_dir):
         polygon_gdf (geo-dataframe): geo-dataframe containing the selected polygons.
 
     Raises:
-        Warning: a warning is raised if the datetime-format of the netCDF-file does not match conventions and/or supported formats.
+        Warning: raised if the datetime-format of the netCDF-file does not match conventions and/or supported formats.
 
     Returns:
         array: filled array containing the variable values (X) and binary conflict data (Y) plus meta-data.
@@ -162,28 +162,42 @@ def fill_XY(XY, config, root_dir, conflict_data, polygon_gdf, out_dir):
 
     return pd.DataFrame.from_dict(XY).to_numpy()
 
-def fill_X_sample(XY, config, root_dir, polygon_gdf, proj_year):
+def fill_X_sample(X, config, root_dir, polygon_gdf, proj_year):
+    """Fills the X-dictionary with the data sample data besides any conflict-related data for each polygon and each year.
+    Used during the projection runs as the sample and conflict data need to be treated separately there.
 
-    # TODO: update this function such that reading conflict at t-1 works with csv stored in previous timestep
+    Args:
+        X (dict): dictionary containing keys to be sampled.
+        config (ConfigParser-object): object containing the parsed configuration-settings of the model.
+        root_dir (str): path to location of cfg-file of reference run.
+        polygon_gdf (geo-dataframe): geo-dataframe containing the selected polygons.
+        proj_year (int): year for which projection is made.
+
+    Raises:
+        Warning: raised if the datetime-format of the netCDF-file does not match conventions and/or supported formats.
+
+    Returns:
+        dict: dictionary containing sample values.
+    """    
 
     print('INFO: entering year {}'.format(proj_year))
 
     # go through all keys in dictionary
-    for key, value in XY.items(): 
+    for key, value in X.items(): 
 
         if key == 'poly_ID':
         
             data_series = value
             data_list = conflict.get_poly_ID(polygon_gdf)
             data_series = data_series.append(pd.Series(data_list), ignore_index=True)
-            XY[key] = data_series
+            X[key] = data_series
 
         elif key == 'poly_geometry':
         
             data_series = value
             data_list = conflict.get_poly_geometry(polygon_gdf, config)
             data_series = data_series.append(pd.Series(data_list), ignore_index=True)
-            XY[key] = data_series
+            X[key] = data_series
 
         else:
 
@@ -195,41 +209,52 @@ def fill_X_sample(XY, config, root_dir, polygon_gdf, proj_year):
                     data_series = value
                     data_list = variables.nc_with_float_timestamp(polygon_gdf, config, root_dir, key, proj_year)
                     data_series = data_series.append(pd.Series(data_list), ignore_index=True)
-                    XY[key] = data_series
+                    X[key] = data_series
                     
                 elif np.dtype(nc_ds.time) == 'datetime64[ns]':
                     data_series = value
                     data_list = variables.nc_with_continous_datetime_timestamp(polygon_gdf, config, root_dir, key, proj_year)
                     data_series = data_series.append(pd.Series(data_list), ignore_index=True)
-                    XY[key] = data_series
+                    X[key] = data_series
                     
                 else:
                     raise Warning('WARNING: this nc-file does have a different dtype for the time variable than currently supported: {}'.format(nc_fo))
 
-    return XY
+    return X
 
-def fill_X_conflict(XY, config, conflict_data, polygon_gdf):
+def fill_X_conflict(X, config, conflict_data, polygon_gdf):
[…]
