### 2021-01-20 79bc1b8
update docs for evaluation.py
