### 2021-01-20 bc7fb51
update docs for evaluation.py
