### 2021-01-20 15b0ffb
updated docs for machine_learning.py
