+++ copro/machine_learning.py
@@ -71,10 +71,6 @@ def split_scale_train_test_split(X, Y, config, scaler):
         config (ConfigParser-object): object containing the parsed configuration-settings of the model.
         scaler (scaler): the specified scaling method instance.
 
-    Raises:
-        AssertionError: raised if after all manipulations the number of unique identifiers does not match number of data points in test-data.
-        AssertionError: raised if after all manipulations the number of geometries does not match number of data points in test-data.
-
     Returns:
         arrays: arrays containing training-data and test-data as well as IDs and geometry for training-data and test-data.
     """ 
@@ -98,43 +94,48 @@ def split_scale_train_test_split(X, Y, config, scaler):
     X_train_ID, X_train_geom, X_train = conflict.split_conflict_geom_data(X_train)
     X_test_ID, X_test_geom, X_test = conflict.split_conflict_geom_data(X_test)
 
-    if not len(X_test_ID) == len(X_test):
-        raise AssertionError('lenght X_test_ID does not match lenght X_test - {} vs {}'.format(len(X_test_ID), len(X_test)))
+    assert (len(X_test_ID) == len(X_test)), AssertionError('ERROR: lenght X_test_ID does not match lenght X_test - {} vs {}'.format(len(X_test_ID), len(X_test)))
 
-    if not len(X_test_geom) == len(X_test):
-        raise AssertionError('lenght X_test_geom does not match lenght X_test - {} vs {}'.format(len(X_test_geom), len(X_test)))
+    assert (len(X_test_geom) == len(X_test)), AssertionError('ERROR: lenght X_test_geom does not match lenght X_test - {} vs {}'.format(len(X_test_geom), len(X_test)))
 
     return X_train, X_test, y_train, y_test, X_train_geom, X_test_geom, X_train_ID, X_test_ID
 
 def fit_predict(X_train, y_train, X_test, clf, config, out_dir, run_nr):
-    """Fits the classifier based on training-data and makes predictions.
+    """Fits classifier based on training-data and makes predictions.
+    Per run_nr, different X_train and y_train data is employed.
+    The fitted classifier is dumped to file with pickle.
     Additionally, the prediction probability is determined.
 
     Args:
-        X_train (array): training-data of variable values
-        y_train ([type]): training-data of conflict data
-        X_test ([type]): test-data of variable values
+        X_train (array): training-data of variable values.
+        y_train (array): training-data of conflict data.
+        X_test (array): test-data of variable values.
         clf (classifier): the specified model instance.
         config (ConfigParser-object): object containing the parsed configuration-settings of the model.
-        out_dir (path): path to output folder
-        run_nr (int): number of fit/predict repetition and created classifier
+        out_dir (path): path to output folder.
+        run_nr (int): number of fit/predict repetition and created classifier.
 
     Returns:
         arrays: arrays including the predictions made and their probabilities
     """    
 
+    # fit the classifier with training data
     clf.fit(X_train, y_train)
 
+    # create folder to store all classifiers with pickle
     clf_pickle_rep = os.path.join(out_dir, 'clfs')
     if not os.path.isdir(clf_pickle_rep):
         os.makedirs(clf_pickle_rep)
 
+    # dump all classifiers
     if config.getboolean('general', 'verbose'): print('DEBUG: dumping classifier to {}'.format(clf_pickle_rep))
     with open(os.path.join(clf_pickle_rep, 'clf_{}.pkl'.format(run_nr)), 'wb') as f:
         pickle.dump(clf, f)
 
+    # make prediction
     y_pred = clf.predict(X_test)
 
+    # make prediction of probability
     y_prob = clf.predict_proba(X_test)
 
     return y_pred, y_prob
[…]
