### 2021-01-20 82a17a7
updated docs for models.py
