+++ copro/models.py
@@ -52,7 +52,7 @@ def leave_one_out(X, Y, config, scaler, clf, out_dir):
         DeprecationWarning: this function will most likely be deprecated due to lack of added value and applicability.
     """    
 
-    raise DeprecationWarning('WARNING: the leave-one-out model will be most likely be deprecated in near future')
+    raise DeprecationWarning('WARNING: the leave-one-out model is not supported anymore and will be deprecated in a future release')
 
     X_train, X_test, y_train, y_test, X_train_geom, X_test_geom, X_train_ID, X_test_ID = machine_learning.split_scale_train_test_split(X, Y, config, scaler)
 
@@ -93,7 +93,7 @@ def single_variables(X, Y, config, scaler, clf, out_dir):
         DeprecationWarning: this function will most likely be deprecated due to lack of added value and applicability.
     """    
 
-    raise DeprecationWarning('WARNING: the single-variable model will be most likely be deprecated in near future')
+    raise DeprecationWarning('WARNING: the single-variable model is not supported anymore and will be deprecated in a future release')
 
     X_train, X_test, y_train, y_test, X_train_geom, X_test_geom, X_train_ID, X_test_ID = machine_learning.split_scale_train_test_split(X, Y, config, scaler)
 
@@ -136,6 +136,7 @@ def dubbelsteen(X, Y, config, scaler, clf, out_dir):
     """   
 
     print('INFO: dubbelsteenmodel running')
+    raise DeprecationWarning('WARNING: the dubbelsteenmodel model is not supported anymore and will be deprecated in a future release')
 
     Y = utils.create_artificial_Y(Y)
 
@@ -152,30 +153,36 @@ def dubbelsteen(X, Y, config, scaler, clf, out_dir):
     return X_df, y_df, eval_dict
 
 def determine_projection_period(config_REF, config_PROJ, out_dir_PROJ):
+    """Determines the period for which projections need to be made. 
+    This is defined as the period between the end year of the reference run and the specified projection year for each projection.
+
+    Args:
+        config_REF (ConfigParser-object): object containing the parsed configuration-settings of the model for the reference run.
+        config_PROJ (ConfigParser-object): object containing the parsed configuration-settings of the model for a projection run..
+        out_dir_PROJ (str): path to output folder of a projection run.
+
+    Returns:
+        list: list containing all years of the projection period.
+    """    
 
     print('INFO: determinining annual conflict occurence from end of reference run until end of projection run')
 
-    # if not os.path.isdir(os.path.join(out_dir_PROJ, 'files')):
-    #     print('DEBUG: creating output folder for annual conflict maps {}'.format(os.path.join(out_dir_PROJ, 'files')))
-    #     os.makedirs(os.path.join(out_dir_PROJ, 'files'))
-
+    # get all years of projection period
     projection_period = np.arange(config_REF.getint('settings', 'y_end')+1, config_PROJ.getint('settings', 'y_proj')+1, 1)
+    # convert to list
     projection_period = projection_period.tolist()
     print('DEBUG: the projection period is {} to {}'.format(projection_period[0], projection_period[-1]))
 
     return projection_period
 
-def predictive(X, clf, scaler, main_dict, root_dir):
+def predictive(X, clf, scaler):
     """Predictive model to use the already fitted classifier to make projections.
     As other models, it reads data which are then scaled and used in conjuction with the classifier to project conflict risk.
 
     Args:
         X (array): array containing the variable values plus unique identifer and geometry information.
-        scaler (scaler): the specified scaling method instance.
-        config (ConfigParser-object): object containing the parsed configuration-settings of the model.
-
-    Raises:
[…]
