### 2021-01-20 d7ed735
updated docs for models.py
