### 2021-01-20 94255ae
improved docs for pipeline.py
