### 2021-01-20 3551484
improved docs for pipeline.py
