+++ example/example_settings.cfg
@@ -34,7 +34,7 @@ type_of_violence=1,2,3
 
 [climate]
 shp=KoeppenGeiger/2000/Koeppen_Geiger_1976-2000.shp
-# define either one or more classes (use abbreviations!) or specify None for not filtering
+# define either one or more classes (use abbreviations!) or specify nothing for not filtering
 zones=BWh,BSh
 code2class=KoeppenGeiger/classification_codes.txt
 
