+++ copro/selection.py
@@ -15,21 +15,26 @@ def filter_conflict_properties(gdf, config):
         geo-dataframe: geo-dataframe containing filtered entries.
     """    
     
+    # create dictinoary with all selection criteria
     selection_criteria = {'best': config.getint('conflict', 'min_nr_casualties'),
                           'type_of_violence': (config.get('conflict', 'type_of_violence')).rsplit(',')}
     
-    print('INFO: filtering on conflict properties.')
+    print('INFO: filtering based on conflict properties.')
     
+    # go through all criteria
     for key in selection_criteria:
 
+        # if nothing is specified, pass
         if selection_criteria[key] == '':
             if config.getboolean('general', 'verbose'): print('DEBUG: passing key', key, 'as it is empty')
             pass
 
+        # for criterion 'best' (i.e. best estimate of fatalities), select all entries above threshold
         elif key == 'best':
             if config.getboolean('general', 'verbose'): print('DEBUG: filtering key', key, 'with lower value', selection_criteria[key])
             gdf = gdf.loc[(gdf[key] >= selection_criteria[key])]
 
+        # for other criteria, select all entries matching the specified value(s) per criterion
         else:
             if config.getboolean('general', 'verbose'): print('DEBUG: filtering key', key, 'with value(s)', selection_criteria[key])
             gdf = gdf.loc[(gdf[key].isin(selection_criteria[key]))]
@@ -47,11 +52,12 @@ def select_period(gdf, config):
         geo-dataframe: geo-dataframe containing filtered entries.
     """    
 
+    # get start and end year of model period
     t0 = config.getint('settings', 'y_start')
     t1 = config.getint('settings', 'y_end')
     
+    # select those entries meeting the requirements
     if config.getboolean('general', 'verbose'): print('DEBUG: focussing on period between {} and {}'.format(t0, t1))
-    
     gdf = gdf.loc[(gdf.year >= t0) & (gdf.year <= t1)]
     
     return gdf
@@ -69,21 +75,27 @@ def clip_to_extent(gdf, config, root_dir):
         geo-dataframe: geo-dataframe containing country polygons of selected continent.
     """    
 
+    # get path to file with polygons for which analysis is carried out
     shp_fo = os.path.join(root_dir, config.get('general', 'input_dir'), config.get('extent', 'shp'))
     
-    print('INFO: reading extent and spatial aggregation level from file {}'.format(shp_fo))
+    # read file
+    if config.getboolean('general', 'verbose'): print('DEBUG: reading extent and spatial aggregation level from file {}'.format(shp_fo))
     extent_gdf = gpd.read_file(shp_fo)
 
-    print('INFO: fixing invalid geometries')
+    # fixing invalid geometries
+    if config.getboolean('general', 'verbose'): print('DEBUG: fixing invalid geometries')
     extent_gdf.geometry = extent_gdf.buffer(0)
 
-    print('INFO: clipping clipping conflict dataset to extent')    
+    # clip the conflict dataframe to the specified polygons
+    if config.getboolean('general', 'verbose'): print('DEBUG: clipping clipping conflict dataset to extent')    
     gdf = gpd.clip(gdf, extent_gdf)
     
     return gdf, extent_gdf
 
 def climate_zoning(gdf, extent_gdf, config, root_dir):
     """This function allows for selecting only those conflicts and polygons falling in specified climate zones.
+    Also, a global dataframe is returned containing the IDs and geometry of all polygons after selection procedure.
+    This can be used to add geometry information to model output based on common ID.
 
     Args:
         gdf (geo-dataframe): geo-dataframe containing conflict data.
[…]
