+++ copro/utils.py
@@ -22,15 +22,17 @@ def get_geodataframe(config, root_dir, longitude='longitude', latitude='latitude
         crs (str, optional): coordinate system to be used for georeferencing. Defaults to 'EPSG:4326'.
 
     Returns:
-        geo-dataframe: ge-referenced conflict data.
+        geo-dataframe: geo-referenced conflict data.
     """     
     
+    # get path to file containing data
     conflict_fo = os.path.join(root_dir, config.get('general', 'input_dir'), config.get('conflict', 'conflict_file'))
 
     # read file to pandas dataframe
     print('INFO: reading csv file to dataframe {}'.format(conflict_fo))
     df = pd.read_csv(conflict_fo)
 
+    # convert dataframe to geo-dataframe
     if config.getboolean('general', 'verbose'): print('DEBUG: translating to geopandas dataframe')
     gdf = gpd.GeoDataFrame(df,
                           geometry=gpd.points_from_xy(df[longitude], df[latitude]),
@@ -56,7 +58,7 @@ def show_versions():
 
     #- Geopandas versions lower than 0.7.0 do not yet have the clip function
     if gpd_version < '0.7.0':
-        sys.exit('please upgrade geopandas to version 0.7.0, your current version is {}'.format(gpd_version))
+        sys.exit('please upgrade geopandas to version 0.7.0, your current version is {}. To avoid the problem, make sure CoPro is installed in its own conda environment.'.format(gpd_version))
 
     print("Python version: {}".format(os_version))
     print("copro version: {}".format(cm_version))
@@ -96,7 +98,7 @@ def parse_projection_settings(config, root_dir):
         config (ConfigParser-object): object containing the parsed configuration-settings of the model for the reference run.
 
     Returns:
-        [dict]: dictionary with name and config-object per specified projection run.
+        dict: dictionary with name and config-object per specified projection run.
     """    
 
     # initiate output dictionary
@@ -112,7 +114,7 @@ def parse_projection_settings(config, root_dir):
         each_val = os.path.abspath(os.path.join(root_dir, each_val))
 
         # parse each config-file specified
-        print('DEBUG: parsing settings from file {}'.format(each_val))
+        if config.getboolean('general', 'verbose'): print('DEBUG: parsing settings from file {}'.format(each_val))
         each_config = parse_settings(each_val)
 
         # update the output dictionary with key and config-object
@@ -121,7 +123,7 @@ def parse_projection_settings(config, root_dir):
     return config_dict
 
 def make_output_dir(config, root_dir, config_dict):
-    """Creates the output folder at location specfied in cfg-file.
+    """Creates the output folder at location specfied in cfg-file, and returns dictionary with config-objects and out-dir per run.
 
     Args:
         config (ConfigParser-object): object containing the parsed configuration-settings of the model.
@@ -132,13 +134,19 @@ def make_output_dir(config, root_dir, config_dict):
         dict: dictionary containing config-objects and output directories for reference run and all projection runs.
     """    
 
+    # get path to main output directory as specified in cfg-file
     out_dir = os.path.join(root_dir, config.get('general','output_dir'))
     print('INFO: saving output to main folder {}'.format(out_dir))
 
-    all_out_dirs = []
+    # initalize list for all out-dirs
+    all_out_dirs = list()
 
+    # append the path to the output folder for the reference run
+    # note that this is hardcoded, i.e. each output folder will have a sub-folder '_REF'
     all_out_dirs.append(os.path.join(out_dir, '_REF'))
 
[…]
