### 2021-01-20 678c01d
updated functions in API description
