+++ docs/api/utils.rst
@@ -11,8 +11,9 @@ Auxiliary functions
    utils.get_geodataframe
    utils.show_versions
    utils.parse_settings
+   utils.parse_projection_settings
    utils.make_output_dir
-   utils.download_PRIO
+   utils.download_UCDP
    utils.initiate_setup
    utils.create_artificial_Y
    utils.global_ID_geom_info
