### 2021-01-20 fe51ea2
updated functions in API description
