+++ docs/api/conflict.rst
@@ -8,6 +8,9 @@ Work with conflict data
    :nosignatures:
 
    conflict.conflict_in_year_bool
+   conflict.conflict_in_previous_year
+   conflict.read_projected_conflict
+   conflict.calc_conflicts_nb
    conflict.get_poly_ID
    conflict.get_poly_geometry
    conflict.split_conflict_geom_data
