### 2021-01-22 #2855 @sbenthall
Good news. I've got the install working. Sorry about that.

Now I'm trying to test the library.
Tiny snag: downloading the demo data and putting it in the right place so that the .cfg files can find it requires a little trial and error for me at the moment.

Certainly not a dealbreaker but some scripting around that would be nice.

I'll continue my review over the weekend -- it should be done by Monday. Thanks for your patience on this--I know it's been a long wait.
