### 2021-01-22 #2855 @JannisHoch
great to hear - i am curious, what exactly was the problem?

so the 'download_example_data.sh' does not what it's supposed to do? guess i'll here that from you soon.
but to not slow down the review process, i could also send you the example data via mail?

looking forward to the remaining review!
