### 2021-01-22 #2855 @sbenthall
> so the 'download_example_data.sh' does not what it's supposed to do?

oh! I didn't see this. It is not mentioned in the README --- it just says the data is on Zenodo. So I fetched it by hand...
