### 2021-01-24 #2855 @sbenthall
https://copro.readthedocs.io/en/latest/examples/nb02_XY_data.ipynb.html
4.2.2.2. "Assing" -> "Assign"
