### 2021-01-25 #2855 @sbenthall
I see your example notebooks use `%matplotlib inline`. I believe more recent versions of Jupyter have made this unnecessary.
