### 2021-01-25 #2855 @sbenthall
As you direct users to use your Jupyter notebook examples, it would be helpful if you referred them to instructions for setting up their Jupyter notebooks to be able to use the conda environment as a kernel. e.g.

https://medium.com/@nrk25693/how-to-add-your-conda-environment-to-your-jupyter-notebook-in-just-4-steps-abeab8b8d084
