### 2021-01-25 #2855 @sbenthall
nb01 cell [5] I'm getting the following error:

`conflict_gdf, extent_gdf, selected_polygons_gdf, global_df = selection.select(config, out_dir, root_dir)`
https://gist.github.com/sbenthall/24da689b34fef8fc2ef9e16b99027215

nb02 cell 3 I'm getting the following error:
`conflict_gdf = gpd.read_file('conflicts.shp')
selected_polygons_gdf = gpd.read_file('polygons.shp')`
https://gist.github.com/sbenthall/d8a722add82fa013d3779e39953773fe

....

I see now that these shapefiles are created by n01 in normal operation, and that the `run_notebooks.sh` script cleans these out. But I'm not deep enough in to know why nb01 is failing when ran manually.

....

I know it's hard to get the tooling around this perfectly but I think with a little more documentation the initial user experience would be more idiot-proof.
