### 2021-01-25 #2855 @JannisHoch
hi @sbenthall ,

thanks for your comments. Regarding the first three, I will change them asap.

Regarding the last, few commetns from my side.

-  seems like you don't have the package `rtree` installed. I checked the `environment.yml` file and it is indeed missing there. but thus far i was under the impression that this was installed along `rasterio` - maybe your initial installation issues can explain this error. either way, i will add `rtree` to the yml-file.
- indeed, some files are written to file in nb01 and then loaded in the other notebooks. this is because when initializing the model, all output files are removed if the output folder already exists. so that is a way to circumvene this 'issue' when using notebooks. note, when using the command line script, this problem of course does not persist!
- the error you get in nb02 is thus related to the one in nb01. if you cannot run nb01 succesfully, the remaining notebooks won't work either since some files are missing!
- i'll work on the documentation in the notebooks asap.
