### 2021-01-25 f97a808
added rtree and ipykernel to requirements
