+++ environment.yml
@@ -10,6 +10,7 @@ dependencies:
   - pandas==1.0.3
   - numpy==1.18.1
   - matplotlib==3.2.1
+  - rtree==0.9.4
   - rasterio==1.1.0
   - pytest>=5.4.2
   - pytest-runner
@@ -18,6 +19,7 @@ dependencies:
   - ipython
   - notebook
   - nbconvert
+  - ipykernel==5.3.2
   - scikit-learn==0.23.2
   - netcdf4
   - ConfigParser
