+++ README.rst
@@ -62,6 +62,12 @@ To install CoPro in editable mode in this environment, run this command next in 
 
     $ pip install -e .
 
+When using Jupyter Notebook, it can be handy to have the copro environment available. It can be installed into Jupyter Notebook with the following command:
+
+.. code-block:: console
+
+    $ python -m ipykernel install --name=copro
+
 Command-line script
 --------------------
 
@@ -108,6 +114,7 @@ Example data
 ----------------
 
 Example data for demonstration purposes can be downloaded from `Zenodo <https://zenodo.org/record/4297295>`_.
+To facilitate this process, the bash-script ``download_example_data.sh`` can be called in the example folder.
 
 With this (or other) data, the provided configuration-files (cfg-files) can be used to perform a reference run or a projection run. 
 All output is stored in the output directory specified in the cfg-files. 
@@ -124,6 +131,8 @@ They can all be run and converted to html-files by executing the provided shell-
     $ sh run_notebooks.sh
 
 It is of course also possible to execute the notebook cell-by-cell and explore the full range of possibilities.
+Note that in this case the notebooks need to be run in the right order as some temporary files will be saved to file in one notebook and loaded in another!
+This is due to the re-initalization of the model at the beginning of each notebook and resulting deletion of all files in existing output folders.
 
 The notebooks are also used to exemplify the `Workflow <https://copro.readthedocs.io/en/latest/examples/index.html>`_ of CoPro.
 
