### 2021-01-25 #2855 @JannisHoch
I have added `rtree` and `ipykernel` to the `environment.yml` file, so hope that a) writing the files in nb01 works now, and b) adding the copro-environmen to your jupyter kernel is possible.

Also, I have fixed the typo, added info about the `download_example_data.sh` file in `README.md`, and placed contextual info on the execution of the notebooks there as well.
