### 2021-01-29 ae00140
removed .pkl suffic for clf folders
