### 2021-01-29 c0c07b1
updated postprocessing script to work with csv
