+++ copro/scripts/postprocessing/plot_polygon_vals.py
@@ -5,21 +5,19 @@ import geopandas as gpd
 import os
 
 @click.command()
-@click.option('-f', '--csv-file',help='path to csv file')
+@click.option('-f', '--file_object',help='path to geojson file_object')
 @click.option('-c', '--column', help='column name')
-@click.option('-t', '--title', help='title for plot and file name')
+@click.option('-t', '--title', help='title for plot and file_object name')
 @click.option('-v0', '--minimum-value')
 @click.option('-v1', '--maximum-value')
 @click.option('-o', '--output-dir', help='path to output directory', type=click.Path())
 
-def main(csv_file=None, column=None, title=None, minimum_value=None, maximum_value=None, output_dir=None):
-    """Quick and dirty function to plot the column values of a shape file with minimum user input, and save plot.
+def main(file_object=None, column=None, title=None, minimum_value=None, maximum_value=None, output_dir=None):
+    """Quick and dirty function to plot the column values of a geojson file with minimum user input, and save plot.
     """
 
-    fo = os.path.abspath(csv_file)
-    click.echo('\nreading csv-file {}'.format(fo))
-    click.echo('... and converting to dataframe')
-    df = pd.Dataframe().from_dict(fo)
+    fo = os.path.abspath(file_object)
+    click.echo('\nreading csv-file_object {}'.format(fo))
     click.echo('... and converting it to geopandas dataframe')
     df = gpd.read_file(fo)
 
@@ -32,11 +30,12 @@ def main(csv_file=None, column=None, title=None, minimum_value=None, maximum_val
             legend=True,
             legend_kwds={'label': str(column),
                          'orientation': "vertical"})
-    ax.set_title(str(title))
+    if title != None:
+        ax.set_title(str(title))
 
-    file_name = 'plt_{0}_fromFile_{1}_andColumn_{2}.png'.format(title, csv_file.rsplit('/')[-1], column)
-    click.echo('saving plot to file {}'.format(os.path.abspath(os.path.join(output_dir, file_name))))
-    plt.savefig(os.path.abspath(os.path.join(output_dir, file_name)), dpi=300, bbox_inches='tight')
+    file_object_name = str(os.path.basename(file_object)).rsplit('.')[0]
+    click.echo('saving plot to file_object {}'.format(os.path.abspath(os.path.join(output_dir, '{}.png'.format(file_object_name)))))
+    plt.savefig(os.path.abspath(os.path.join(output_dir, '{}.png'.format(file_object_name))), dpi=300, bbox_inches='tight')
 
 if __name__ == '__main__':
 
