### 2021-02-01 #2855 @sbenthall
Ok, just tried the `.sh` scripts version of working README.

I ran `sh run_script_reference.sh` no problem.

Then with `sh run_script_projections.sh` I got this error:

https://gist.github.com/sbenthall/11fdd17af6f690c13933383e504eac50
