### 2021-02-02 #2855 @JannisHoch
hi @sbenthall,
I am again stligthly puzzled about the errror you get.
If I run the scripts in the 'joss_review' branch (which is the one under review I guess?), I don't get any issues at all.
Also, I get different console output than you.

sure you use the right branch?
