### 2021-02-03 8888bbc
changed colormap to rgb
