### 2021-02-03 59084e9
plotting dev of output over time
