### 2021-02-03 866973f
converting geojsons to one GIF
