### 2021-02-05 629a1c9
option added to not delete png output
