### 2021-02-10 ebf680d
added assertions
