+++ copro/pipeline.py
@@ -2,6 +2,7 @@ from copro import models, data, machine_learning, evaluation, utils
 import pandas as pd
 import numpy as np
 import pickle
+import click
 import os, sys
 
 
@@ -33,13 +34,13 @@ def create_XY(config, out_dir, root_dir, polygon_gdf, conflict_gdf):
         XY = data.fill_XY(XY, config, root_dir, conflict_gdf, polygon_gdf, out_dir)
 
         # save array to XY.npy out_dir
-        print('INFO: saving XY data by default to file {}'.format(os.path.join(out_dir, 'XY.npy')))
+        click.echo('INFO: saving XY data by default to file {}'.format(os.path.join(out_dir, 'XY.npy')))
         np.save(os.path.join(out_dir,'XY'), XY)
 
     # if path to XY.npy is specified, read the data intead
     else:
 
-        print('INFO: loading XY data from file {}'.format(os.path.join(root_dir, config.get('pre_calc', 'XY'))))
+        click.echo('INFO: loading XY data from file {}'.format(os.path.join(root_dir, config.get('pre_calc', 'XY'))))
         XY = np.load(os.path.join(root_dir, config.get('pre_calc', 'XY')), allow_pickle=True)
         
     # split the XY data into sample data X and target values Y
@@ -131,10 +132,10 @@ def run_prediction(scaler, main_dict, root_dir, selected_polygons_gdf):
     for (each_key, each_val) in config_REF.items('PROJ_files'):
 
         # get config-object and out-dir per projection
-        print('INFO: loading config-object for projection run: {}'.format(each_key))
+        click.echo('INFO: loading config-object for projection run: {}'.format(each_key))
         config_PROJ = main_dict[str(each_key)][0][0]
         out_dir_PROJ = main_dict[str(each_key)][1]
-        print('DEBUG: storing output for this projections to folder {}'.format(out_dir_PROJ))
+        click.echo('DEBUG: storing output for this projections to folder {}'.format(out_dir_PROJ))
 
         # if not os.path.isdir(os.path.join(out_dir_PROJ, 'files')):
         #     os.makedirs(os.path.join(out_dir_PROJ, 'files'))
@@ -149,18 +150,18 @@ def run_prediction(scaler, main_dict, root_dir, selected_polygons_gdf):
         for i in range(len(projection_period)):
 
             proj_year = projection_period[i]
-            print('INFO: making projection for year {}'.format(proj_year))
+            click.echo('INFO: making projection for year {}'.format(proj_year))
 
             X = data.initiate_X_data(config_PROJ)
-            print('INFO: reading sample data from files')
+            click.echo('INFO: reading sample data from files')
             X = data.fill_X_sample(X, config_PROJ, root_dir, selected_polygons_gdf, proj_year)
 
             # for the first projection year, we need to fall back on the observed conflict at the last time step of the reference run
             if i == 0:
-                print('INFO: reading previous conflicts from file {}'.format(os.path.join(out_dir_REF, 'files', 'conflicts_in_{}.csv'.format(config_REF.getint('settings', 'y_end')))))
+                click.echo('INFO: reading previous conflicts from file {}'.format(os.path.join(out_dir_REF, 'files', 'conflicts_in_{}.csv'.format(config_REF.getint('settings', 'y_end')))))
                 conflict_data = pd.read_csv(os.path.join(out_dir_REF, 'files', 'conflicts_in_{}.csv'.format(config_REF.getint('settings', 'y_end'))), index_col=0)
 
-                print('DEBUG: combining sample data with conflict data from previous year')
+                click.echo('DEBUG: combining sample data with conflict data from previous year')
                 X = data.fill_X_conflict(X, config_PROJ, conflict_data, selected_polygons_gdf)
[…]
