### 2021-02-11 7e18ca3
working on assertion
