### 2021-02-11 c9411e4
rstats now considers all touched cells
