### 2021-02-11 eb9cb04
improved click.echo statements
