### 2021-02-11 cfa5ccc
added ID output if -inf is calculated
