### 2021-02-11 a4bdc39
accounting for NaNs
