+++ copro/variables.py
@@ -84,7 +84,9 @@ def nc_with_float_timestamp(extent_gdf, config, root_dir, var_name, sim_year):
         prov = extent_gdf.iloc[i]
 
         # compute zonal stats for this province
-        zonal_stats = rstats.zonal_stats(prov.geometry, nc_arr_vals, affine=affine, stats=stat_method)
+        # computes a value per polygon for all raster cells that are touched by polygon (all_touched=True)
+        # if all_touched=False, only for raster cells with centre point in polygon are considered, but this is problematic for very small polygons
+        zonal_stats = rstats.zonal_stats(prov.geometry, nc_arr_vals, affine=affine, stats=stat_method, all_touched=True)
         val = zonal_stats[0][stat_method]
 
         # # if specified, log-transform value
@@ -104,8 +106,6 @@ def nc_with_float_timestamp(extent_gdf, config, root_dir, var_name, sim_year):
             click.echo('WARNING: NaN computed!')
 
         list_out.append(val)
-
-    if config.getboolean('general', 'verbose'): click.echo('DEBUG: ... done.')
 
     return list_out
 
@@ -200,12 +200,10 @@ def nc_with_continous_datetime_timestamp(extent_gdf, config, root_dir, var_name,
             else:
                 val = val_ln
 
-        # click.echo a warning if result is None
-        if (val == None) and (config.getboolean('general', 'verbose')): 
-            click.echo('WARNING: NaN computed!')
-
+        # print a warning if result is None
+        if (val == None) or (val == np.nan) and (config.getboolean('general', 'verbose')): 
+            click.echo('WARNING: {} computed for ID {}!'.format(val, prov.watprovID))
+        
         list_out.append(val)
-
-    if config.getboolean('general', 'verbose'): click.echo('DEBUG: ... done.')
 
     return list_out
