### 2021-02-11 0628b99
accounting for NaNs
