### 2021-02-11 04c42fc
make avg over time predictions
