+++ copro/scripts/postprocessing/avg_over_time.py
@@ -1,0 +1,110 @@
+import click
+import glob
+import numpy as np
+import geopandas as gpd
+import pandas as pd
+import matplotlib.pyplot as plt
+from copro import utils, evaluation
+import os, sys
+
+@click.command()
+@click.option('-t0', '--start-year', type=int)
+@click.option('-t1', '--end-year', type=int)
+@click.option('--geojson/--no-geojson', help='save output to geojson or not', default=False)
+@click.option('--png/--no-png', help='save output to png or not', default=True)
+@click.argument('input_dir', type=click.Path())
+@click.argument('output_dir',type=click.Path())
+@click.argument('selected_polygons', type=click.Path())
+
+def main(input_dir=None, output_dir=None, selected_polygons=None, start_year=None, end_year=None, geojson=None, png=None):
+    """Post-processing script to calculate average model output over a user-specifeid period or all output geojson-files.
+    """    
+
+    # check if start/end time settings are consistent
+    if ((start_year != None) and (end_year == None)) or ((end_year != None) and (start_year == None)):
+        raise ValueError('ERROR: if start or end year is specified, the pendant must be specified too!')
+
+    # read a shp-file with geometries of all selected polygons
+    click.echo('\nreading shp-file with all polygons from {}'.format(os.path.abspath(selected_polygons)))
+    selected_polygons_gdf = gpd.read_file(os.path.abspath(selected_polygons))
+    # create dataframe
+    global_df = utils.global_ID_geom_info(selected_polygons_gdf)
+
+    # find all geojson-files in input-dir
+    input_dir = os.path.abspath(input_dir)
+    click.echo('getting geojson-files from {}'.format(input_dir))
+    all_files = sorted(glob.glob(os.path.join(input_dir, '*.geojson')))
+
+    # if specific start/end time is specified, find only those geojson-files for specified period
+    if (start_year != None) and (end_year != None):
+        # define period between start and ent time
+        period = np.arange(start_year, end_year+1, 1)
+        click.echo('using all geojson-files for years {} to {}'.format(period[0], period[-1]))
+        # creating suffix for file saving later
+        suffix = '{}_to_{}'.format(period[0], period[-1])
+        # initiate empty list for selected geojson-files
+        selected_files = []
+        # select
+        for fo in all_files:
+            # if the year-suffix of geojson-file matches year in period, add to list
+            year = int(str(str(os.path.basename(fo)).rsplit('.')[0]).rsplit('_')[-1])
+            if year in period:
+                print('adding to selection')
+                selected_files.append(fo)
+    # if not end/start time is specified, use all geojson-files in input-dir
+    else:
+        click.echo('using all geojson-files in input-dir')
+        # also here, create suffix for file saving laster
+        suffix = 'all_years'
+        selected_files = all_files
+
+    # initiatie empyt dataframe for collecting all annual output
+    y_df = pd.DataFrame()
+
+    # go through all geojson-files left over after selection step
+    for geojson in selected_files:
+        # read files and convert to datatrame
+        click.echo('reading file {}'.format(geojson))
+        gdf = gpd.read_file(geojson, driver='GeoJSON')
+        df = pd.DataFrame(gdf)
+        # append to dataframe 
+        y_df = y_df.append(df, ignore_index=True)
+    
+    # initiate dataframe for time-averaged output
+    click.echo('creating one output dataframe from all geojson-files')
+    y_out = pd.DataFrame()
+    # get all unique IDs of polygons
[…]
