+++ copro/pipeline.py
@@ -192,6 +192,9 @@ def run_prediction(scaler, main_dict, root_dir, selected_polygons_gdf):
 
                 X = pd.DataFrame(X)
                 if config_REF.getboolean('general', 'verbose'): click.echo('DEBUG: number of data points including missing values: {}'.format(len(X)))
+
+                # print(X[X.isna().any(axis=1)])
+
                 X = X.dropna()
                 if config_REF.getboolean('general', 'verbose'): click.echo('DEBUG: number of data points excluding missing values: {}'.format(len(X)))
                 
