+++ example/example_settings.cfg
@@ -23,7 +23,7 @@ testZWEI=./example_settings_proj.cfg
 XY=
 
 [extent]
-shp=waterProvinces/waterProvinces_Africa.shp
+shp=waterProvinces/waterProvinces_Africa_eliminatedPolysLE20000km2.shp
 
 [conflict]
 # either specify path to file or state 'download' to download latest PRIO/UCDP dataset
