### 2021-02-11 78d7841
added debugging about NaNs; fillna ipv dropna
