### 2021-02-11 8c6ffde
added debugging about NaNs; fillna ipv dropna
