### 2021-02-12 #2855 @JannisHoch
dear @sbenthall ,
hope you are doing well.
I am just curious how the review process is going on your side.
Not wanna be pushy as I know that this work is voluntary and Covid-19 is not doing any good either...
Since I want to use the code in a manuscript, I'd very much appreciate it if there could be a citable version of the code until then.
Thanks!
Looking forward to any new bugs you find.
Jannis
