### 2021-02-16 #2855 @sbenthall
I'm using the `joss_review` branch and am getting the same error.

What is your output for that script?
