### 2021-02-16 #2855 @JannisHoch
that's great news, congrats!

i am bit puzzled by your error message, or more precisely by the console output you get.
it does not look like anything i get when using the `joss_review` branch with the cfg-files as is plus the downloaded example data.
it seems like you getting debug output, but not the one i get using the `joss_review` branch.
that is what makes me assume that there is something odd about your model exeuction, but cannot pinpoint what it is.

did you alter any of the files by any chance?
maybe good if you could send me the cfg-files you use if the error persists?
