### 2021-02-17 eb94755
included makedirs() for output dir
