### 2021-02-18 #2855 @JannisHoch
@sbenthall so what was the magic trick? anything i can add to the package to make the installation bullet-proof?
