### 2021-02-18 #2855 @sbenthall
It was likely just human error initially. Don't worry about it.

Next steps:
- proofread the paper and make sure there are no typos.
- Double check the MISSING DOIs that @whedon brought up

This is the final editing stage.
