### 2021-02-18 #2855 @sbenthall
If you have edits to the paper, then yes, that's right.

When it's absolutely finalized, you tag it.
