### 2021-02-18 #2855 @sbenthall
Yeah, those DOI's Whedon pointed out should be added to the joss .bib file
