### 2021-02-18 8c9da5a
updated paper, refs, and copro version nr
