### 2021-02-18 42bfeda
updated paper, refs, and copro version nr
