### 2021-02-18 #2855 @sbenthall
"a key concept in mind when establishing CoPro was to be able to produce reproducible output using transparent models"

grammar issue ... "a concept" "was" "to be able to"

suggested revision: "CoPro produces reproducible output using transparent models."
