### 2021-02-18 #2855 @sbenthall
I'm wondering about the title "a data-driven model for conflict risk projections" .
From the article text, it sounds like CoPro is a toolkit allowing for many different models?
