### 2021-02-18 #2855 @JannisHoch
this depends how one defines model, a maybe bit philosophical question.
so i'd say that CoPro is a model in itself, but makes use of other models again... so maybe call it a modelling framework?
