### 2021-02-18 #2855 @JannisHoch
the version number is 0.0.7 and the zenodo DOI is https://doi.org/10.5281/zenodo.4548904.
