### 2021-02-18 #2855 @JannisHoch
that's how it's supposed to work? or do I need to let wheedon update version number and DOI? not fully clear to me from the instructions i must admit.
