### 2021-02-18 #2855 @sbenthall
That's weird. Maybe whedon needs to regenerate the PDF before checking references again
