### 2021-02-19 #2855 @arfon
@sbenthall @JannisHoch - I've updated the submission with the Zenodo archive. Next steps here are:

- Run `@whedon accept from branch joss_review `

Or, have @JannisHoch merge their changes into the default branch on their repository which means you then would need to run:

- `@whedon accept`

In both cases, the `@whedon accept` command will produce the final PDF proofs and notify the EiC on rotation that this submission is ready to accepted and published.
