### 2021-02-19 #2855 @JannisHoch
@arfon
many thanks for fixing this.

@arfon @sbenthall
I have merged branch `joss_review` into default (which is `dev`, or does whedon choosing `master` as default by default?).
